[SYCLomatic][NFC] Move rules for Random/Solver/Blas/Spblas API into RulesMathLib folder.
