+++ clang/lib/DPCT/RulesLang/RulesLang.cpp
@@ -2603,145 +2603,6 @@ void CU_JITEnumsRule::runRule(const MatchFinder::MatchResult &Result) {
 }
 
 
-// Rule for BLAS enums.
-// Migrate BLAS status values to corresponding int values
-// Other BLAS named values are migrated to corresponding named values
-void BLASEnumsRule::registerMatcher(MatchFinder &MF) {
-  MF.addMatcher(declRefExpr(to(enumConstantDecl(matchesName(
-                                "(CUBLAS_STATUS.*)|(CUDA_R_.*)|(CUDA_C_.*)|("
-                                "CUBLAS_GEMM_.*)|(CUBLAS_POINTER_MODE.*)"))))
-                    .bind("BLASStatusConstants"),
-                this);
-  MF.addMatcher(
-      declRefExpr(to(enumConstantDecl(matchesName(
-                      "(CUBLAS_OP.*)|(CUBLAS_SIDE.*)|(CUBLAS_FILL_"
-                      "MODE.*)|(CUBLAS_DIAG.*)|(CUBLAS_.*_MATH)|CUBLAS_MATH_"
-                      "DISALLOW_REDUCED_PRECISION_REDUCTION|(CUBLAS_COMPUTE_.*)"
-                      "|(CUBLASLT_ORDER_.*)"
-                      "|(CUBLASLT_POINTER_MODE_.*)|(CUBLASLT_MATRIX_LAYOUT_.*)|"
-                      "(CUBLASLT_MATMUL_DESC_.*)|(CUBLASLT_MATRIX_TRANSFORM_"
-                      "DESC_.*)|(CUBLASLT_EPILOGUE_.*)"))))
-          .bind("BLASNamedValueConstants"),
-      this);
-}
-
-void BLASEnumsRule::runRule(const MatchFinder::MatchResult &Result) {
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "BLASStatusConstants")) {
-    auto *EC = cast<EnumConstantDecl>(DE->getDecl());
-    emplaceTransformation(new ReplaceStmt(DE, toString(EC->getInitVal(), 10)));
-  }
-
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "BLASNamedValueConstants")) {
-    auto *EC = cast<EnumConstantDecl>(DE->getDecl());
-    std::string Name = EC->getNameAsString();
-    auto Search = MapNames::BLASEnumsMap.find(Name);
-    if (Search == MapNames::BLASEnumsMap.end()) {
-      llvm::dbgs() << "[" << getName()
-                   << "] Unexpected enum variable: " << Name;
-      return;
-    }
-    std::string Replacement = Search->second;
-    emplaceTransformation(new ReplaceStmt(DE, std::move(Replacement)));
-  }
-}
-
-
-// Rule for RANDOM enums.
-void RandomEnumsRule::registerMatcher(MatchFinder &MF) {
-  MF.addMatcher(
-      declRefExpr(to(enumConstantDecl(matchesName("CURAND_STATUS.*"))))
-          .bind("RANDOMStatusConstants"),
-      this);
-  MF.addMatcher(
-      declRefExpr(to(enumConstantDecl(matchesName("CURAND_ORDERING.*"))))
-          .bind("RANDOMOrderingConstants"),
-      this);
-  MF.addMatcher(declRefExpr(to(enumConstantDecl(matchesName("CURAND_RNG.*"))))
-                    .bind("RandomTypeEnum"),
-                this);
-}
-
-void RandomEnumsRule::runRule(const MatchFinder::MatchResult &Result) {
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "RANDOMStatusConstants")) {
-    auto *EC = cast<EnumConstantDecl>(DE->getDecl());
-    emplaceTransformation(new ReplaceStmt(DE, toString(EC->getInitVal(), 10)));
-  }
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "RANDOMOrderingConstants")) {
-    std::string EnumStr = DE->getNameInfo().getName().getAsString();
-    auto Search = MapNames::RandomOrderingTypeMap.find(EnumStr);
-    if (Search == MapNames::RandomOrderingTypeMap.end()) {
-      report(DE->getBeginLoc(), Diagnostics::API_NOT_MIGRATED, false, EnumStr);
-      return;
-    }
-    emplaceTransformation(new ReplaceStmt(DE, Search->second));
-  }
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "RandomTypeEnum")) {
-    std::string EnumStr = DE->getNameInfo().getName().getAsString();
-    auto Search = MapNames::RandomEngineTypeMap.find(EnumStr);
-    if (Search == MapNames::RandomEngineTypeMap.end()) {
-      report(DE->getBeginLoc(), Diagnostics::API_NOT_MIGRATED, false, EnumStr);
-      return;
-    }
-    if (EnumStr == "CURAND_RNG_PSEUDO_XORWOW" ||
-        EnumStr == "CURAND_RNG_QUASI_SOBOL64" ||
-        EnumStr == "CURAND_RNG_QUASI_SCRAMBLED_SOBOL64") {
-      report(DE->getBeginLoc(), Diagnostics::DIFFERENT_GENERATOR, false);
-    } else if (EnumStr == "CURAND_RNG_QUASI_SCRAMBLED_SOBOL32") {
-      report(DE->getBeginLoc(), Diagnostics::DIFFERENT_BASIC_GENERATOR, false);
-    }
-    emplaceTransformation(new ReplaceStmt(DE, Search->second));
-  }
-}
-
-
-// Rule for spBLAS enums.
-// Migrate spBLAS status values to corresponding int values
-// Other spBLAS named values are migrated to corresponding named values
-void SPBLASEnumsRule::registerMatcher(MatchFinder &MF) {
-  MF.addMatcher(declRefExpr(to(enumConstantDecl(matchesName(
-                                "(CUSPARSE_STATUS.*)|(CUSPARSE_POINTER_MODE.*)|"
-                                "(CUSPARSE_ALG.*)|(CUSPARSE_SOLVE_POLICY.*)|("
-                                "CUSPARSE_SPSM_ALG_.*)"))))
-                    .bind("SPBLASStatusConstants"),
-                this);
-  MF.addMatcher(
-      declRefExpr(
-          to(enumConstantDecl(matchesName(
-              "(CUSPARSE_OPERATION_.*)|(CUSPARSE_FILL_MODE_.*)|("
-              "CUSPARSE_DIAG_TYPE_.*)|(CUSPARSE_INDEX_.*)|(CUSPARSE_"
-              "MATRIX_TYPE_.*)|(CUSPARSE_ORDER_.*)|(CUSPARSE_ACTION_.*)"))))
-          .bind("SPBLASNamedValueConstants"),
-      this);
-}
-
-void SPBLASEnumsRule::runRule(const MatchFinder::MatchResult &Result) {
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "SPBLASStatusConstants")) {
-    auto *EC = cast<EnumConstantDecl>(DE->getDecl());
-    emplaceTransformation(new ReplaceStmt(DE, toString(EC->getInitVal(), 10)));
-  }
-
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "SPBLASNamedValueConstants")) {
-    auto *EC = cast<EnumConstantDecl>(DE->getDecl());
-    std::string Name = EC->getNameAsString();
-    auto Search = MapNames::SPBLASEnumsMap.find(Name);
-    if (Search == MapNames::SPBLASEnumsMap.end()) {
-      llvm::dbgs() << "[" << getName()
-                   << "] Unexpected enum variable: " << Name;
-      return;
-    }
-    std::string Replacement = Search->second;
-    emplaceTransformation(new ReplaceStmt(DE, std::move(Replacement)));
-  }
-}
-
-
 /// The function returns the migrated arguments of the scalar parameters.
 /// In the original code, the type of this parameter is pointer.
 /// (1) If original type is float/double and argument is like "&alpha",
@@ -2783,2170 +2644,7 @@ std::string getValueStr(const Expr *Expr, std::string ExprStr,
          QueueStr + ")";
 }
 
-// Rule for spBLAS function calls.
-void SPBLASFunctionCallRule::registerMatcher(MatchFinder &MF) {
-  auto functionName = [&]() {
-    return hasAnyName(
-        /*management*/
-        "cusparseCreate", "cusparseDestroy", "cusparseSetStream",
-        "cusparseGetStream", "cusparseGetPointerMode", "cusparseSetPointerMode",
-        "cusparseGetErrorName", "cusparseGetErrorString", "cusparseGetProperty",
-        /*helper*/
-        "cusparseCreateMatDescr", "cusparseDestroyMatDescr",
-        "cusparseSetMatType", "cusparseGetMatType", "cusparseSetMatIndexBase",
-        "cusparseGetMatIndexBase", "cusparseSetMatDiagType",
-        "cusparseGetMatDiagType", "cusparseSetMatFillMode",
-        "cusparseGetMatFillMode", "cusparseCreateSolveAnalysisInfo",
-        "cusparseDestroySolveAnalysisInfo", "cusparseCreateCsrsv2Info",
-        "cusparseDestroyCsrsv2Info",
-        /*level 2*/
-        "cusparseScsrmv", "cusparseDcsrmv", "cusparseCcsrmv", "cusparseZcsrmv",
-        "cusparseScsrmv_mp", "cusparseDcsrmv_mp", "cusparseCcsrmv_mp",
-        "cusparseZcsrmv_mp", "cusparseCsrmvEx", "cusparseCsrmvEx_bufferSize",
-        "cusparseScsrsv_analysis", "cusparseDcsrsv_analysis",
-        "cusparseCcsrsv_analysis", "cusparseZcsrsv_analysis",
-        "cusparseScsrsv_solve", "cusparseDcsrsv_solve", "cusparseCcsrsv_solve",
-        "cusparseZcsrsv_solve", "cusparseScsrsv2_bufferSize",
-        "cusparseDcsrsv2_bufferSize", "cusparseCcsrsv2_bufferSize",
-        "cusparseZcsrsv2_bufferSize", "cusparseScsrsv2_analysis",
-        "cusparseDcsrsv2_analysis", "cusparseCcsrsv2_analysis",
-        "cusparseZcsrsv2_analysis", "cusparseScsrsv2_solve",
-        "cusparseDcsrsv2_solve", "cusparseCcsrsv2_solve",
-        "cusparseZcsrsv2_solve", "cusparseCcsrsv2_bufferSizeExt",
-        "cusparseDcsrsv2_bufferSizeExt", "cusparseScsrsv2_bufferSizeExt",
-        "cusparseZcsrsv2_bufferSizeExt", "cusparseCsrsv_analysisEx",
-        "cusparseCsrsv_solveEx",
-        /*level 3*/
-        "cusparseScsrmm", "cusparseDcsrmm", "cusparseCcsrmm", "cusparseZcsrmm",
-        "cusparseScsrgemm", "cusparseDcsrgemm", "cusparseCcsrgemm",
-        "cusparseZcsrgemm", "cusparseXcsrgemmNnz", "cusparseScsrmm2",
-        "cusparseDcsrmm2", "cusparseCcsrmm2", "cusparseZcsrmm2",
-        /*Generic*/
-        "cusparseCreateCsr", "cusparseDestroySpMat", "cusparseCsrGet",
-        "cusparseSpMatGetFormat", "cusparseSpMatGetIndexBase",
-        "cusparseSpMatGetValues", "cusparseSpMatSetValues",
-        "cusparseCreateDnMat", "cusparseDestroyDnMat", "cusparseDnMatGet",
-        "cusparseDnMatGetValues", "cusparseDnMatSetValues",
-        "cusparseCreateDnVec", "cusparseDestroyDnVec", "cusparseDnVecGet",
-        "cusparseDnVecGetValues", "cusparseDnVecSetValues",
-        "cusparseCsrSetPointers", "cusparseSpMatGetSize",
-        "cusparseSpMatGetAttribute", "cusparseSpMatSetAttribute",
-        "cusparseCreateConstDnVec", "cusparseConstDnVecGet",
-        "cusparseConstDnVecGetValues", "cusparseSpMM",
-        "cusparseSpMM_bufferSize", "cusparseSpMV", "cusparseSpMV_bufferSize",
-        "cusparseSpMM_preprocess", "cusparseSpGEMM_compute",
-        "cusparseSpGEMM_copy", "cusparseSpGEMM_createDescr",
-        "cusparseSpGEMM_destroyDescr", "cusparseSpGEMM_workEstimation",
-        "cusparseScsr2csc", "cusparseDcsr2csc", "cusparseCcsr2csc",
-        "cusparseZcsr2csc", "cusparseCsr2cscEx2_bufferSize",
-        "cusparseCsr2cscEx2", "cusparseSpSV_createDescr",
-        "cusparseSpSV_destroyDescr", "cusparseSpSV_solve",
-        "cusparseSpSV_bufferSize", "cusparseSpSV_analysis",
-        "cusparseSpSM_analysis", "cusparseSpSM_bufferSize",
-        "cusparseSpSM_createDescr", "cusparseSpSM_destroyDescr",
-        "cusparseSpSM_solve", "cusparseCreateCoo");
-  };
-  MF.addMatcher(
-      callExpr(allOf(callee(functionDecl(functionName())), parentStmt()))
-          .bind("FunctionCall"),
-      this);
-  MF.addMatcher(callExpr(allOf(callee(functionDecl(functionName())),
-                               unless(parentStmt())))
-                    .bind("FunctionCallUsed"),
-                this);
-}
 
-void SPBLASFunctionCallRule::runRule(const MatchFinder::MatchResult &Result) {
-  const CallExpr *CE = getNodeAsType<CallExpr>(Result, "FunctionCall");
-  if (!CE) {
-    if (!(CE = getNodeAsType<CallExpr>(Result, "FunctionCallUsed")))
-      return;
-  }
-
-  if (!CE->getDirectCallee())
-    return;
-
-  std::string FuncName =
-      CE->getDirectCallee()->getNameInfo().getName().getAsString();
-  StringRef FuncNameRef(FuncName);
-  if (FuncNameRef.ends_with("csrmv") || FuncNameRef.ends_with("csrmv_mp")) {
-    report(
-        DpctGlobalInfo::getSourceManager().getExpansionLoc(CE->getBeginLoc()),
-        Diagnostics::UNSUPPORT_MATRIX_TYPE, true,
-        "general/symmetric/triangular");
-  } else if (FuncNameRef.ends_with("csrmm") ||
-             FuncNameRef.ends_with("csrmm2")) {
-    report(
-        DpctGlobalInfo::getSourceManager().getExpansionLoc(CE->getBeginLoc()),
-        Diagnostics::UNSUPPORT_MATRIX_TYPE, true, "general");
-  }
-
-  if (CallExprRewriterFactoryBase::RewriterMap->find(FuncName) !=
-      CallExprRewriterFactoryBase::RewriterMap->end()) {
-    ExprAnalysis EA(CE);
-    emplaceTransformation(EA.getReplacement());
-    EA.applyAllSubExprRepl();
-    return;
-  }
-  if (FuncName == "cusparseXcsrgemmNnz") {
-    std::vector<std::string> MigratedArgs;
-    for (const auto &Arg : CE->arguments()) {
-      MigratedArgs.push_back(ExprAnalysis::ref(Arg));
-    }
-    // We need find the next cusparse<T>csrgemm API call which is using the
-    // result of this API call, otherwise a warning will be emitted.
-    auto findOuterCS = [](const Stmt *Input) {
-      const CompoundStmt *CS = nullptr;
-      DpctGlobalInfo::findAncestor<Stmt>(
-          Input, [&](const DynTypedNode &Cur) -> bool {
-            if (Cur.get<DoStmt>() || Cur.get<ForStmt>() ||
-                Cur.get<WhileStmt>() || Cur.get<SwitchStmt>() ||
-                Cur.get<IfStmt>())
-              return true;
-            if (const CompoundStmt *S = Cur.get<CompoundStmt>())
-              CS = S;
-            return false;
-          });
-      return CS;
-    };
-    const CompoundStmt *CS1 = findOuterCS(CE);
-    // Find all the cusparse<T>csrgemm calls in this range.
-    using namespace clang::ast_matchers;
-    auto Matcher =
-        findAll(callExpr(callee(functionDecl(hasAnyName(
-                             "cusparseScsrgemm", "cusparseDcsrgemm",
-                             "cusparseCcsrgemm", "cusparseZcsrgemm"))))
-                    .bind("CallExpr"));
-    auto CEResults = match(Matcher, *CS1, DpctGlobalInfo::getContext());
-    // Find the correct call
-    const CallExpr* CorrectCall = nullptr;
-    for (auto &Result : CEResults) {
-      const CallExpr *MatchedCE = Result.getNodeAs<CallExpr>("CallExpr");
-      if (MatchedCE) {
-        // 1. The context should be the same
-        const CompoundStmt *CS2 = findOuterCS(MatchedCE);
-        if (CS1 != CS2)
-          continue;
-        // 2. The args should be the same
-        std::vector<std::string> MatchedCEMigratedArgs;
-        for (const auto &Arg : MatchedCE->arguments()) {
-          MatchedCEMigratedArgs.push_back(ExprAnalysis::ref(Arg));
-        }
-        if ([&]() -> bool {
-              const static std::map<unsigned /*CE*/, unsigned /*MatchedCE*/>
-                  IdxMap = {
-                      {0, 0},   {1, 1},   {2, 2},   {3, 3},
-                      {4, 4},   {5, 5},   {6, 6},   {7, 7},
-                      {8, 9},   {9, 10},  {10, 11}, {11, 12},
-                      {12, 14}, {13, 15}, {14, 16}, {15, 18},
-                  };
-              for (const auto &P : IdxMap) {
-                if (MigratedArgs[P.first] != MatchedCEMigratedArgs[P.second]) {
-                  return false;
-                }
-              }
-              return true;
-            }()) {
-          CorrectCall = MatchedCE;
-          break;
-        }
-      }
-    }
-    const constexpr int Placeholder = -1;
-    std::map<int /*CE*/, int /*MatchedCE*/> InsertBeforeIdxMap;
-    if (CorrectCall) {
-      InsertBeforeIdxMap = {
-          {8, 8},
-          {12, 13},
-      };
-    } else {
-      report(
-          DpctGlobalInfo::getSourceManager().getExpansionLoc(CE->getBeginLoc()),
-          Diagnostics::SPARSE_NNZ, true);
-      InsertBeforeIdxMap = {
-          {8, Placeholder},
-          {12, Placeholder},
-      };
-    }
-    std::string MigratedCall;
-    MigratedCall = MapNames::getDpctNamespace() + "sparse::csrgemm_nnz(";
-    for (unsigned i = 0; i < MigratedArgs.size(); i++) {
-      if (auto Iter = InsertBeforeIdxMap.find(i);
-          Iter != InsertBeforeIdxMap.end()) {
-        if (Iter->second == Placeholder) {
-          MigratedCall += ("dpct_placeholder, ");
-        } else {
-          MigratedCall += (ExprAnalysis::ref(
-                               CorrectCall->getArg(InsertBeforeIdxMap.at(i))) +
-                           ", ");
-        }
-      }
-      MigratedCall += MigratedArgs[i];
-      if (i != MigratedArgs.size() - 1)
-        MigratedCall += ", ";
-    }
-    MigratedCall += ")";
-    auto DefRange = getDefinitionRange(CE->getBeginLoc(), CE->getEndLoc());
-    SourceLocation Begin = DefRange.getBegin();
-    SourceLocation End = DefRange.getEnd();
-    End = End.getLocWithOffset(
-        Lexer::MeasureTokenLength(End, DpctGlobalInfo::getSourceManager(),
-                                  DpctGlobalInfo::getContext().getLangOpts()));
-    emplaceTransformation(replaceText(Begin, End, std::move(MigratedCall),
-                                      DpctGlobalInfo::getSourceManager()));
-    return;
-  }
-}
-
-
-// Rule for Random function calls. Currently only support host APIs.
-void RandomFunctionCallRule::registerMatcher(MatchFinder &MF) {
-  auto functionName = [&]() {
-    return hasAnyName(
-        "curandCreateGenerator", "curandSetPseudoRandomGeneratorSeed",
-        "curandSetGeneratorOffset", "curandSetQuasiRandomGeneratorDimensions",
-        "curandDestroyGenerator", "curandGenerate", "curandGenerateLongLong",
-        "curandGenerateLogNormal", "curandGenerateLogNormalDouble",
-        "curandGenerateNormal", "curandGenerateNormalDouble",
-        "curandGeneratePoisson", "curandGenerateUniform",
-        "curandGenerateUniformDouble", "curandSetStream",
-        "curandCreateGeneratorHost", "curandSetGeneratorOrdering");
-  };
-  MF.addMatcher(
-      callExpr(allOf(callee(functionDecl(functionName())), parentStmt()))
-          .bind("FunctionCall"),
-      this);
-  MF.addMatcher(callExpr(allOf(callee(functionDecl(functionName())),
-                               unless(parentStmt())))
-                    .bind("FunctionCallUsed"),
-                this);
-}
-
-void RandomFunctionCallRule::runRule(const MatchFinder::MatchResult &Result) {
-  bool IsAssigned = false;
-  const CallExpr *CE = getNodeAsType<CallExpr>(Result, "FunctionCall");
-  if (!CE) {
-    if (!(CE = getNodeAsType<CallExpr>(Result, "FunctionCallUsed")))
-      return;
-    IsAssigned = true;
-  }
-
-  if (!CE->getDirectCallee())
-    return;
-
-  auto &SM = DpctGlobalInfo::getSourceManager();
-
-  std::string FuncName =
-      CE->getDirectCallee()->getNameInfo().getName().getAsString();
-  SourceLocation FuncNameBegin(CE->getBeginLoc());
-  SourceLocation FuncCallEnd(CE->getEndLoc());
-  // Below code can distinguish this kind of function like macro
-  //  #define CHECK_STATUS(x) x
-  //  CHECK_STATUS(anAPICall());
-  bool IsMacroArg = SM.isMacroArgExpansion(CE->getBeginLoc()) &&
-                    SM.isMacroArgExpansion(CE->getEndLoc());
-
-  if (FuncNameBegin.isMacroID() && IsMacroArg) {
-    FuncNameBegin = SM.getImmediateSpellingLoc(FuncNameBegin);
-    FuncNameBegin = SM.getExpansionLoc(FuncNameBegin);
-  } else if (FuncNameBegin.isMacroID()) {
-    FuncNameBegin = SM.getExpansionLoc(FuncNameBegin);
-  }
-
-  if (FuncCallEnd.isMacroID() && IsMacroArg) {
-    FuncCallEnd = SM.getImmediateSpellingLoc(FuncCallEnd);
-    FuncCallEnd = SM.getExpansionLoc(FuncCallEnd);
-  } else if (FuncCallEnd.isMacroID()) {
-    FuncCallEnd = SM.getExpansionLoc(FuncCallEnd);
-  }
-
-  // Offset 1 is the length of the last token ")"
-  FuncCallEnd = SM.getExpansionLoc(FuncCallEnd).getLocWithOffset(1);
-
-  if (IsAssigned) {
-    requestFeature(HelperFeatureEnum::device_ext);
-    insertAroundStmt(CE, MapNames::getCheckErrorMacroName() + "(", ")");
-  }
-
-  if (FuncName == "curandCreateGenerator" ||
-      FuncName == "curandCreateGeneratorHost") {
-    const auto *const Arg0 = CE->getArg(0);
-    requestFeature(HelperFeatureEnum::device_ext);
-    std::string RHS =
-        buildString(" = ", MapNames::getLibraryHelperNamespace(),
-                    "rng::create_host_rng(", ExprAnalysis::ref(CE->getArg(1)));
-    if (FuncName == "curandCreateGeneratorHost") {
-      if (DpctGlobalInfo::useSYCLCompat())
-        RHS = buildString(RHS, ", *", MapNames::getDpctNamespace(),
-                          "cpu_device().default_queue()");
-      else
-        RHS = buildString(RHS, ", ", MapNames::getDpctNamespace(),
-                          "cpu_device().default_queue()");
-    }
-    RHS = buildString(RHS, ")");
-    if (Arg0->getStmtClass() == Stmt::UnaryOperatorClass) {
-      const auto *const UO = cast<const UnaryOperator>(Arg0);
-      auto SE = UO->getSubExpr();
-      if (UO->getOpcode() == UO_AddrOf &&
-          (SE->getStmtClass() == Stmt::DeclRefExprClass ||
-           SE->getStmtClass() == Stmt::MemberExprClass)) {
-        return emplaceTransformation(new ReplaceStmt(
-            CE, false, buildString(ExprAnalysis::ref(SE), RHS)));
-      }
-    }
-    return emplaceTransformation(new ReplaceStmt(
-        CE, false,
-        buildString("*(", ExprAnalysis::ref(CE->getArg(0)), ")", RHS)));
-  }
-  if (FuncName == "curandDestroyGenerator") {
-    return emplaceTransformation(new ReplaceStmt(
-        CE, false, buildString(ExprAnalysis::ref(CE->getArg(0)), ".reset()")));
-  }
-  if (FuncName == "curandSetPseudoRandomGeneratorSeed") {
-    return emplaceTransformation(new ReplaceStmt(
-        CE, false,
-        buildString(ExprAnalysis::ref(CE->getArg(0)), "->set_seed(",
-                    ExprAnalysis::ref(CE->getArg(1)), ")")));
-  }
-  if (FuncName == "curandSetQuasiRandomGeneratorDimensions") {
-    return emplaceTransformation(new ReplaceStmt(
-        CE, false,
-        buildString(ExprAnalysis::ref(CE->getArg(0)), "->set_dimensions(",
-                    ExprAnalysis::ref(CE->getArg(1)), ")")));
-  }
-  if (MapNames::RandomGenerateFuncMap.find(FuncName) !=
-      MapNames::RandomGenerateFuncMap.end()) {
-    auto ArgStr = ExprAnalysis::ref(CE->getArg(1));
-    for (unsigned i = 2; i < CE->getNumArgs(); ++i) {
-      ArgStr += buildString(", ", ExprAnalysis::ref(CE->getArg(i)));
-    }
-    return emplaceTransformation(new ReplaceStmt(
-        CE, false,
-        buildString(
-            ExprAnalysis::ref(CE->getArg(0)),
-            "->" + MapNames::RandomGenerateFuncMap.find(FuncName)->second + "(",
-            ArgStr, ")")));
-  }
-  if (FuncName == "curandSetGeneratorOffset") {
-    return emplaceTransformation(new ReplaceStmt(
-        CE, false,
-        buildString(ExprAnalysis::ref(CE->getArg(0)), "->skip_ahead(",
-                    ExprAnalysis::ref(CE->getArg(1)), ")")));
-  }
-  if (FuncName == "curandSetStream") {
-    return emplaceTransformation(new ReplaceStmt(
-        CE, false,
-        buildString(ExprAnalysis::ref(CE->getArg(0)), "->set_queue(",
-                    ExprAnalysis::ref(CE->getArg(1)), ")")));
-  }
-  ExprAnalysis EA(CE);
-  emplaceTransformation(EA.getReplacement());
-  EA.applyAllSubExprRepl();
-}
-
-// Rule for device Random function calls.
-void DeviceRandomFunctionCallRule::registerMatcher(MatchFinder &MF) {
-  auto functionName = [&]() {
-    return hasAnyName(
-        "curand_init", "curand", "curand4", "curand_normal", "curand_normal4",
-        "curand_normal2", "curand_normal2_double", "curand_normal_double",
-        "curand_log_normal", "curand_log_normal2", "curand_log_normal2_double",
-        "curand_log_normal4", "curand_log_normal_double", "curand_uniform",
-        "curand_uniform2_double", "curand_uniform4", "curand_uniform_double",
-        "curand_poisson", "curand_poisson4", "skipahead", "skipahead_sequence",
-        "skipahead_subsequence", "curand_uniform4_double", "curand_normal4_double",
-        "curand_log_normal4_double");
-  };
-  MF.addMatcher(
-      callExpr(callee(functionDecl(functionName()))).bind("FunctionCall"),
-      this);
-}
-
-void DeviceRandomFunctionCallRule::runRule(
-    const MatchFinder::MatchResult &Result) {
-  const CallExpr *CE = getNodeAsType<CallExpr>(Result, "FunctionCall");
-  if (!CE)
-    return;
-  if (!CE->getDirectCallee())
-    return;
-
-  auto &SM = DpctGlobalInfo::getSourceManager();
-  auto SL = SM.getExpansionLoc(CE->getBeginLoc());
-  std::string Key =
-      SM.getFilename(SL).str() + std::to_string(SM.getDecomposedLoc(SL).second);
-  DpctGlobalInfo::updateInitSuffixIndexInRule(
-      DpctGlobalInfo::getSuffixIndexInitValue(Key));
-
-  std::string FuncName =
-      CE->getDirectCallee()->getNameInfo().getName().getAsString();
-  SourceLocation FuncNameBegin(CE->getBeginLoc());
-  SourceLocation FuncCallEnd(CE->getEndLoc());
-  // Offset 1 is the length of the last token ")"
-  FuncCallEnd = SM.getExpansionLoc(FuncCallEnd).getLocWithOffset(1);
-  auto FuncCallLength =
-      SM.getCharacterData(FuncCallEnd) - SM.getCharacterData(FuncNameBegin);
-  std::string IndentStr = getIndent(FuncNameBegin, SM).str();
-
-  if (FuncName == "curand_init") {
-    if (CE->getNumArgs() < 4) {
-      report(FuncNameBegin, Diagnostics::API_NOT_MIGRATED, false, FuncName);
-      return;
-    }
-
-    std::string Arg0Type = DpctGlobalInfo::getTypeName(
-        CE->getArg(0)->getType().getCanonicalType());
-    std::string Arg1Type = DpctGlobalInfo::getTypeName(
-        CE->getArg(1)->getType().getCanonicalType());
-    std::string Arg2Type = DpctGlobalInfo::getTypeName(
-        CE->getArg(2)->getType().getCanonicalType());
-    std::string DRefArg3Type;
-
-    if (Arg0Type == "unsigned long long" && Arg1Type == "unsigned long long" &&
-        Arg2Type == "unsigned long long" &&
-        CE->getArg(3)->getType().getCanonicalType()->isPointerType()) {
-      DRefArg3Type = DpctGlobalInfo::getTypeName(
-          CE->getArg(3)->getType().getCanonicalType()->getPointeeType());
-      if (MapNames::DeviceRandomGeneratorTypeMap.find(DRefArg3Type) ==
-          MapNames::DeviceRandomGeneratorTypeMap.end()) {
-        report(FuncNameBegin, Diagnostics::NOT_SUPPORTED_PARAMETER, false,
-               FuncName,
-               "parameter " + getStmtSpelling(CE->getArg(3)) +
-                   " is unsupported");
-        return;
-      }
-    } else {
-      report(FuncNameBegin, Diagnostics::API_NOT_MIGRATED, false, FuncName);
-      return;
-    }
-
-    auto IsLiteral = [=](const Expr *E) {
-      if (dyn_cast<IntegerLiteral>(E->IgnoreCasts()) ||
-          dyn_cast<FloatingLiteral>(E->IgnoreCasts()) ||
-          dyn_cast<FixedPointLiteral>(E->IgnoreCasts())) {
-        return true;
-      }
-      return false;
-    };
-
-    std::string GeneratorType =
-        MapNames::DeviceRandomGeneratorTypeMap.find(DRefArg3Type)->second;
-    std::string RNGSeed = ExprAnalysis::ref(CE->getArg(0));
-    bool IsRNGSubseqLiteral = IsLiteral(CE->getArg(1));
-    std::string RNGSubseq = ExprAnalysis::ref(CE->getArg(1));
-    bool IsRNGOffsetLiteral = IsLiteral(CE->getArg(2));
-    std::string RNGOffset = ExprAnalysis::ref(CE->getArg(2));
-    std::string RNGStateName = getDrefName(CE->getArg(3));
-
-    std::string FirstOffsetArg, SecondOffsetArg;
-    if (IsRNGOffsetLiteral) {
-      FirstOffsetArg = RNGOffset;
-    } else {
-      FirstOffsetArg = "static_cast<std::uint64_t>(" + RNGOffset + ")";
-    }
-
-    std::string ReplStr;
-    if (DRefArg3Type == "curandStateXORWOW") {
-      report(FuncNameBegin, Diagnostics::SUBSEQUENCE_IGNORED, false, RNGSubseq);
-      ReplStr = RNGStateName + " = " + GeneratorType + "(" + RNGSeed + ", " +
-                FirstOffsetArg + ")";
-    } else {
-      std::string Factor = "8";
-      if (GeneratorType == MapNames::getLibraryHelperNamespace() +
-                               "rng::device::rng_generator<oneapi::"
-                               "mkl::rng::device::philox4x32x10<1>>" &&
-          DRefArg3Type == "curandStatePhilox4_32_10") {
-        Factor = "4";
-      }
-
-      if (needExtraParens(CE->getArg(1))) {
-        RNGSubseq = "(" + RNGSubseq + ")";
-      }
-      if (IsRNGSubseqLiteral) {
-        SecondOffsetArg = RNGSubseq + " * " + Factor;
-      } else {
-        SecondOffsetArg =
-            "static_cast<std::uint64_t>(" + RNGSubseq + " * " + Factor + ")";
-      }
-
-      ReplStr = RNGStateName + " = " + GeneratorType + "(" + RNGSeed + ", {" +
-                FirstOffsetArg + ", " + SecondOffsetArg + "})";
-    }
-    emplaceTransformation(
-        new ReplaceText(FuncNameBegin, FuncCallLength, std::move(ReplStr)));
-  } else if (FuncName == "skipahead" || FuncName == "skipahead_sequence" ||
-             FuncName == "skipahead_subsequence") {
-    if (FuncName == "skipahead") {
-      std::string Arg1Type = CE->getArg(1)->getType().getAsString();
-      if (Arg1Type != "curandStateMRG32k3a_t *" &&
-          Arg1Type != "curandStatePhilox4_32_10_t *" &&
-          Arg1Type != "curandStateXORWOW_t *") {
-        // Do not support Sobol32 state and Sobol64 state
-        report(FuncNameBegin, Diagnostics::API_NOT_MIGRATED, false, FuncName);
-        return;
-      }
-    }
-    ExprAnalysis EA(CE);
-    emplaceTransformation(EA.getReplacement());
-    EA.applyAllSubExprRepl();
-  } else {
-    ExprAnalysis EA(CE);
-    emplaceTransformation(EA.getReplacement());
-    EA.applyAllSubExprRepl();
-  }
-}
-
-void BLASFunctionCallRule::registerMatcher(MatchFinder &MF) {
-  auto functionName = [&]() {
-    return hasAnyName(
-        /*Regular BLAS API*/
-        /*Regular helper*/
-        "cublasCreate_v2", "cublasDestroy_v2", "cublasSetVector",
-        "cublasGetVector", "cublasSetVectorAsync", "cublasGetVectorAsync",
-        "cublasSetMatrix", "cublasGetMatrix", "cublasSetMatrixAsync",
-        "cublasGetMatrixAsync", "cublasSetStream_v2", "cublasGetStream_v2",
-        "cublasGetPointerMode_v2", "cublasSetPointerMode_v2",
-        "cublasGetAtomicsMode", "cublasSetAtomicsMode", "cublasGetVersion_v2",
-        "cublasGetMathMode", "cublasSetMathMode", "cublasGetStatusString",
-        "cublasSetWorkspace_v2",
-        /*Regular level 1*/
-        "cublasIsamax_v2", "cublasIdamax_v2", "cublasIcamax_v2",
-        "cublasIzamax_v2", "cublasIsamin_v2", "cublasIdamin_v2",
-        "cublasIcamin_v2", "cublasIzamin_v2", "cublasSasum_v2",
-        "cublasDasum_v2", "cublasScasum_v2", "cublasDzasum_v2",
-        "cublasSaxpy_v2", "cublasDaxpy_v2", "cublasCaxpy_v2", "cublasZaxpy_v2",
-        "cublasScopy_v2", "cublasDcopy_v2", "cublasCcopy_v2", "cublasZcopy_v2",
-        "cublasSdot_v2", "cublasDdot_v2", "cublasCdotu_v2", "cublasCdotc_v2",
-        "cublasZdotu_v2", "cublasZdotc_v2", "cublasSnrm2_v2", "cublasDnrm2_v2",
-        "cublasScnrm2_v2", "cublasDznrm2_v2", "cublasSrot_v2", "cublasDrot_v2",
-        "cublasCsrot_v2", "cublasZdrot_v2", "cublasCrot_v2", "cublasZrot_v2",
-        "cublasSrotg_v2", "cublasDrotg_v2", "cublasCrotg_v2", "cublasZrotg_v2",
-        "cublasSrotm_v2", "cublasDrotm_v2", "cublasSrotmg_v2",
-        "cublasDrotmg_v2", "cublasSscal_v2", "cublasDscal_v2", "cublasCscal_v2",
-        "cublasCsscal_v2", "cublasZscal_v2", "cublasZdscal_v2",
-        "cublasSswap_v2", "cublasDswap_v2", "cublasCswap_v2", "cublasZswap_v2",
-        /*Regular level 2*/
-        "cublasSgbmv_v2", "cublasDgbmv_v2", "cublasCgbmv_v2", "cublasZgbmv_v2",
-        "cublasSgemv_v2", "cublasDgemv_v2", "cublasCgemv_v2", "cublasZgemv_v2",
-        "cublasSger_v2", "cublasDger_v2", "cublasCgeru_v2", "cublasCgerc_v2",
-        "cublasZgeru_v2", "cublasZgerc_v2", "cublasSsbmv_v2", "cublasDsbmv_v2",
-        "cublasSspmv_v2", "cublasDspmv_v2", "cublasSspr_v2", "cublasDspr_v2",
-        "cublasSspr2_v2", "cublasDspr2_v2", "cublasSsymv_v2", "cublasDsymv_v2",
-        "cublasCsymv_v2", "cublasZsymv_v2", "cublasSsyr_v2", "cublasDsyr_v2",
-        "cublasSsyr2_v2", "cublasDsyr2_v2", "cublasCsyr_v2", "cublasZsyr_v2",
-        "cublasCsyr2_v2", "cublasZsyr2_v2", "cublasStbmv_v2", "cublasDtbmv_v2",
-        "cublasCtbmv_v2", "cublasZtbmv_v2", "cublasStbsv_v2", "cublasDtbsv_v2",
-        "cublasCtbsv_v2", "cublasZtbsv_v2", "cublasStpmv_v2", "cublasDtpmv_v2",
-        "cublasCtpmv_v2", "cublasZtpmv_v2", "cublasStpsv_v2", "cublasDtpsv_v2",
-        "cublasCtpsv_v2", "cublasZtpsv_v2", "cublasStrmv_v2", "cublasDtrmv_v2",
-        "cublasCtrmv_v2", "cublasZtrmv_v2", "cublasStrsv_v2", "cublasDtrsv_v2",
-        "cublasCtrsv_v2", "cublasZtrsv_v2", "cublasChemv_v2", "cublasZhemv_v2",
-        "cublasChbmv_v2", "cublasZhbmv_v2", "cublasChpmv_v2", "cublasZhpmv_v2",
-        "cublasCher_v2", "cublasZher_v2", "cublasCher2_v2", "cublasZher2_v2",
-        "cublasChpr_v2", "cublasZhpr_v2", "cublasChpr2_v2", "cublasZhpr2_v2",
-        /*Regular level 3*/
-        "cublasSgemm_v2", "cublasDgemm_v2", "cublasCgemm_v2", "cublasZgemm_v2",
-        "cublasHgemm", "cublasCgemm3m", "cublasZgemm3m",
-        "cublasHgemmStridedBatched", "cublasSgemmStridedBatched",
-        "cublasDgemmStridedBatched", "cublasCgemmStridedBatched",
-        "cublasZgemmStridedBatched", "cublasSsymm_v2", "cublasDsymm_v2",
-        "cublasCsymm_v2", "cublasZsymm_v2", "cublasSsyrk_v2", "cublasDsyrk_v2",
-        "cublasCsyrk_v2", "cublasZsyrk_v2", "cublasSsyr2k_v2",
-        "cublasDsyr2k_v2", "cublasCsyr2k_v2", "cublasZsyr2k_v2",
-        "cublasStrsm_v2", "cublasDtrsm_v2", "cublasCtrsm_v2", "cublasZtrsm_v2",
-        "cublasChemm_v2", "cublasZhemm_v2", "cublasCherk_v2", "cublasZherk_v2",
-        "cublasCher2k_v2", "cublasZher2k_v2", "cublasSsyrkx", "cublasDsyrkx",
-        "cublasCsyrkx", "cublasZsyrkx", "cublasCherkx", "cublasZherkx",
-        "cublasStrmm_v2", "cublasDtrmm_v2", "cublasCtrmm_v2", "cublasZtrmm_v2",
-        "cublasHgemmBatched", "cublasSgemmBatched", "cublasDgemmBatched",
-        "cublasCgemmBatched", "cublasZgemmBatched", "cublasStrsmBatched",
-        "cublasDtrsmBatched", "cublasCtrsmBatched", "cublasZtrsmBatched",
-        /*Extensions*/
-        "cublasSgetrfBatched", "cublasDgetrfBatched", "cublasCgetrfBatched",
-        "cublasZgetrfBatched", "cublasSgetrsBatched", "cublasDgetrsBatched",
-        "cublasCgetrsBatched", "cublasZgetrsBatched", "cublasSgetriBatched",
-        "cublasDgetriBatched", "cublasCgetriBatched", "cublasZgetriBatched",
-        "cublasSgeqrfBatched", "cublasDgeqrfBatched", "cublasCgeqrfBatched",
-        "cublasZgeqrfBatched", "cublasSgelsBatched", "cublasDgelsBatched",
-        "cublasCgelsBatched", "cublasZgelsBatched", "cublasGemmEx",
-        "cublasSgemmEx", "cublasCgemmEx", "cublasCgemm3mEx", "cublasNrm2Ex",
-        "cublasDotEx", "cublasDotcEx", "cublasScalEx", "cublasAxpyEx",
-        "cublasRotEx", "cublasGemmBatchedEx", "cublasGemmStridedBatchedEx",
-        "cublasSdgmm", "cublasDdgmm", "cublasCdgmm", "cublasZdgmm",
-        "cublasSgeam", "cublasDgeam", "cublasCgeam", "cublasZgeam",
-        "cublasCopyEx", "cublasSwapEx", "cublasIamaxEx", "cublasIaminEx",
-        "cublasAsumEx", "cublasRotmEx", "cublasCsyrkEx", "cublasCsyrk3mEx",
-        "cublasCherkEx", "cublasCherk3mEx",
-        /*Legacy API*/
-        "cublasInit", "cublasShutdown", "cublasGetError",
-        "cublasSetKernelStream", "cublasGetVersion",
-        /*level 1*/
-        "cublasSnrm2", "cublasDnrm2", "cublasScnrm2", "cublasDznrm2",
-        "cublasSdot", "cublasDdot", "cublasCdotu", "cublasCdotc", "cublasZdotu",
-        "cublasZdotc", "cublasSscal", "cublasDscal", "cublasCscal",
-        "cublasZscal", "cublasCsscal", "cublasZdscal", "cublasSaxpy",
-        "cublasDaxpy", "cublasCaxpy", "cublasZaxpy", "cublasScopy",
-        "cublasDcopy", "cublasCcopy", "cublasZcopy", "cublasSswap",
-        "cublasDswap", "cublasCswap", "cublasZswap", "cublasIsamax",
-        "cublasIdamax", "cublasIcamax", "cublasIzamax", "cublasIsamin",
-        "cublasIdamin", "cublasIcamin", "cublasIzamin", "cublasSasum",
-        "cublasDasum", "cublasScasum", "cublasDzasum", "cublasSrot",
-        "cublasDrot", "cublasCsrot", "cublasZdrot", "cublasCrot", "cublasZrot",
-        "cublasSrotg", "cublasDrotg", "cublasSrotm", "cublasDrotm",
-        "cublasSrotmg", "cublasDrotmg",
-        /*level 2*/
-        "cublasSgemv", "cublasDgemv", "cublasCgemv", "cublasZgemv",
-        "cublasSgbmv", "cublasDgbmv", "cublasCgbmv", "cublasZgbmv",
-        "cublasStrmv", "cublasDtrmv", "cublasCtrmv", "cublasZtrmv",
-        "cublasStbmv", "cublasDtbmv", "cublasCtbmv", "cublasZtbmv",
-        "cublasStpmv", "cublasDtpmv", "cublasCtpmv", "cublasZtpmv",
-        "cublasStrsv", "cublasDtrsv", "cublasCtrsv", "cublasZtrsv",
-        "cublasStpsv", "cublasDtpsv", "cublasCtpsv", "cublasZtpsv",
-        "cublasStbsv", "cublasDtbsv", "cublasCtbsv", "cublasZtbsv",
-        "cublasSsymv", "cublasDsymv", "cublasChemv", "cublasZhemv",
-        "cublasSsbmv", "cublasDsbmv", "cublasChbmv", "cublasZhbmv",
-        "cublasSspmv", "cublasDspmv", "cublasChpmv", "cublasZhpmv",
-        "cublasSger", "cublasDger", "cublasCgeru", "cublasCgerc", "cublasZgeru",
-        "cublasZgerc", "cublasSsyr", "cublasDsyr", "cublasCher", "cublasZher",
-        "cublasSspr", "cublasDspr", "cublasChpr", "cublasZhpr", "cublasSsyr2",
-        "cublasDsyr2", "cublasCher2", "cublasZher2", "cublasSspr2",
-        "cublasDspr2", "cublasChpr2", "cublasZhpr2",
-        /*level 3*/
-        "cublasSgemm", "cublasDgemm", "cublasCgemm", "cublasZgemm",
-        "cublasSsyrk", "cublasDsyrk", "cublasCsyrk", "cublasZsyrk",
-        "cublasCherk", "cublasZherk", "cublasSsyr2k", "cublasDsyr2k",
-        "cublasCsyr2k", "cublasZsyr2k", "cublasCher2k", "cublasZher2k",
-        "cublasSsymm", "cublasDsymm", "cublasCsymm", "cublasZsymm",
-        "cublasChemm", "cublasZhemm", "cublasStrsm", "cublasDtrsm",
-        "cublasCtrsm", "cublasZtrsm", "cublasStrmm", "cublasDtrmm",
-        "cublasCtrmm", "cublasZtrmm",
-        /*64-bit API*/
-        /*helper*/
-        "cublasSetVector_64", "cublasGetVector_64", "cublasSetMatrix_64",
-        "cublasGetMatrix_64", "cublasSetVectorAsync_64",
-        "cublasGetVectorAsync_64", "cublasSetMatrixAsync_64",
-        "cublasGetMatrixAsync_64",
-        /*level 1*/
-        "cublasIsamax_v2_64", "cublasIdamax_v2_64", "cublasIcamax_v2_64",
-        "cublasIzamax_v2_64", "cublasIsamin_v2_64", "cublasIdamin_v2_64",
-        "cublasIcamin_v2_64", "cublasIzamin_v2_64", "cublasSnrm2_v2_64",
-        "cublasDnrm2_v2_64", "cublasScnrm2_v2_64", "cublasDznrm2_v2_64",
-        "cublasSdot_v2_64", "cublasDdot_v2_64", "cublasCdotu_v2_64",
-        "cublasCdotc_v2_64", "cublasZdotu_v2_64", "cublasZdotc_v2_64",
-        "cublasSscal_v2_64", "cublasDscal_v2_64", "cublasCscal_v2_64",
-        "cublasCsscal_v2_64", "cublasZscal_v2_64", "cublasZdscal_v2_64",
-        "cublasSaxpy_v2_64", "cublasDaxpy_v2_64", "cublasCaxpy_v2_64",
-        "cublasZaxpy_v2_64", "cublasScopy_v2_64", "cublasDcopy_v2_64",
-        "cublasCcopy_v2_64", "cublasZcopy_v2_64", "cublasSswap_v2_64",
-        "cublasDswap_v2_64", "cublasCswap_v2_64", "cublasZswap_v2_64",
-        "cublasSasum_v2_64", "cublasDasum_v2_64", "cublasScasum_v2_64",
-        "cublasDzasum_v2_64", "cublasSrot_v2_64", "cublasDrot_v2_64",
-        "cublasCrot_v2_64", "cublasCsrot_v2_64", "cublasZrot_v2_64",
-        "cublasZdrot_v2_64", "cublasSrotm_v2_64", "cublasDrotm_v2_64",
-        /*level 2*/
-        "cublasSgemv_v2_64", "cublasDgemv_v2_64", "cublasCgemv_v2_64",
-        "cublasZgemv_v2_64", "cublasSgbmv_v2_64", "cublasDgbmv_v2_64",
-        "cublasCgbmv_v2_64", "cublasZgbmv_v2_64", "cublasStrmv_v2_64",
-        "cublasDtrmv_v2_64", "cublasCtrmv_v2_64", "cublasZtrmv_v2_64",
-        "cublasStbmv_v2_64", "cublasDtbmv_v2_64", "cublasCtbmv_v2_64",
-        "cublasZtbmv_v2_64", "cublasStpmv_v2_64", "cublasDtpmv_v2_64",
-        "cublasCtpmv_v2_64", "cublasZtpmv_v2_64", "cublasStrsv_v2_64",
-        "cublasDtrsv_v2_64", "cublasCtrsv_v2_64", "cublasZtrsv_v2_64",
-        "cublasStpsv_v2_64", "cublasDtpsv_v2_64", "cublasCtpsv_v2_64",
-        "cublasZtpsv_v2_64", "cublasStbsv_v2_64", "cublasDtbsv_v2_64",
-        "cublasCtbsv_v2_64", "cublasZtbsv_v2_64", "cublasSsymv_v2_64",
-        "cublasDsymv_v2_64", "cublasCsymv_v2_64", "cublasZsymv_v2_64",
-        "cublasChemv_v2_64", "cublasZhemv_v2_64", "cublasSsbmv_v2_64",
-        "cublasDsbmv_v2_64", "cublasChbmv_v2_64", "cublasZhbmv_v2_64",
-        "cublasSspmv_v2_64", "cublasDspmv_v2_64", "cublasChpmv_v2_64",
-        "cublasZhpmv_v2_64", "cublasSger_v2_64", "cublasDger_v2_64",
-        "cublasCgeru_v2_64", "cublasCgerc_v2_64", "cublasZgeru_v2_64",
-        "cublasZgerc_v2_64", "cublasSsyr_v2_64", "cublasDsyr_v2_64",
-        "cublasCsyr_v2_64", "cublasZsyr_v2_64", "cublasCher_v2_64",
-        "cublasZher_v2_64", "cublasSspr_v2_64", "cublasDspr_v2_64",
-        "cublasChpr_v2_64", "cublasZhpr_v2_64", "cublasSsyr2_v2_64",
-        "cublasDsyr2_v2_64", "cublasCsyr2_v2_64", "cublasZsyr2_v2_64",
-        "cublasCher2_v2_64", "cublasZher2_v2_64", "cublasSspr2_v2_64",
-        "cublasDspr2_v2_64", "cublasChpr2_v2_64", "cublasZhpr2_v2_64",
-        /*level 3*/
-        "cublasSgemm_v2_64", "cublasDgemm_v2_64", "cublasCgemm_v2_64",
-        "cublasZgemm_v2_64", "cublasSsyrk_v2_64", "cublasDsyrk_v2_64",
-        "cublasCsyrk_v2_64", "cublasZsyrk_v2_64", "cublasSsymm_v2_64",
-        "cublasDsymm_v2_64", "cublasCsymm_v2_64", "cublasZsymm_v2_64",
-        "cublasStrsm_v2_64", "cublasDtrsm_v2_64", "cublasCtrsm_v2_64",
-        "cublasZtrsm_v2_64", "cublasChemm_v2_64", "cublasZhemm_v2_64",
-        "cublasCherk_v2_64", "cublasZherk_v2_64", "cublasSsyr2k_v2_64",
-        "cublasDsyr2k_v2_64", "cublasCsyr2k_v2_64", "cublasZsyr2k_v2_64",
-        "cublasCher2k_v2_64", "cublasZher2k_v2_64", "cublasSgeam_64",
-        "cublasDgeam_64", "cublasCgeam_64", "cublasZgeam_64", "cublasSdgmm_64",
-        "cublasDdgmm_64", "cublasCdgmm_64", "cublasZdgmm_64",
-        "cublasStrmm_v2_64", "cublasDtrmm_v2_64", "cublasCtrmm_v2_64",
-        "cublasZtrmm_v2_64", "cublasSsyrkx_64", "cublasDsyrkx_64",
-        "cublasCsyrkx_64", "cublasZsyrkx_64", "cublasCherkx_64",
-        "cublasZherkx_64", "cublasHgemm_64", "cublasCgemm3m_64",
-        "cublasZgemm3m_64",
-        /*extension*/
-        "cublasNrm2Ex_64", "cublasDotEx_64", "cublasDotcEx_64",
-        "cublasScalEx_64", "cublasAxpyEx_64", "cublasRotEx_64",
-        "cublasGemmBatchedEx_64", "cublasGemmStridedBatchedEx_64",
-        "cublasCopyEx_64", "cublasSwapEx_64", "cublasIamaxEx_64",
-        "cublasIaminEx_64", "cublasAsumEx_64", "cublasRotmEx_64",
-        "cublasSgemmEx_64", "cublasCgemmEx_64", "cublasCgemm3mEx_64",
-        "cublasGemmEx_64", "cublasCsyrkEx_64", "cublasCsyrk3mEx_64",
-        "cublasCherkEx_64", "cublasCherk3mEx_64",
-        /*cublasLt*/
-        "cublasLtCreate", "cublasLtDestroy", "cublasLtMatmulDescCreate",
-        "cublasLtMatmulDescDestroy", "cublasLtMatmulDescSetAttribute",
-        "cublasLtMatmulDescGetAttribute", "cublasLtMatrixLayoutCreate",
-        "cublasLtMatrixLayoutDestroy", "cublasLtMatrixLayoutGetAttribute",
-        "cublasLtMatrixLayoutSetAttribute", "cublasLtMatmul",
-        "cublasLtMatmulPreferenceCreate", "cublasLtMatmulPreferenceDestroy",
-        "cublasLtMatmulPreferenceSetAttribute",
-        "cublasLtMatmulPreferenceGetAttribute",
-        "cublasLtMatmulAlgoGetHeuristic", "cublasLtMatrixTransformDescCreate",
-        "cublasLtMatrixTransformDescDestroy",
-        "cublasLtMatrixTransformDescSetAttribute",
-        "cublasLtMatrixTransformDescGetAttribute", "cublasLtMatrixTransform",
-        "cublasLtGetVersion");
-  };
-
-  MF.addMatcher(callExpr(allOf(callee(functionDecl(functionName())),
-                               hasAncestor(functionDecl(
-                                   anyOf(hasAttr(attr::CUDADevice),
-                                         hasAttr(attr::CUDAGlobal))))))
-                    .bind("kernelCall"),
-                this);
-
-  MF.addMatcher(
-      callExpr(
-          allOf(callee(functionDecl(functionName())), parentStmt(),
-                hasAncestor(functionDecl(unless(allOf(
-                    hasAttr(attr::CUDADevice), hasAttr(attr::CUDAGlobal)))))))
-          .bind("FunctionCall"),
-      this);
-  MF.addMatcher(
-      callExpr(
-          allOf(callee(functionDecl(functionName())), unless(parentStmt()),
-                unless(hasAncestor(varDecl())),
-                hasAncestor(functionDecl(unless(allOf(
-                    hasAttr(attr::CUDADevice), hasAttr(attr::CUDAGlobal)))))))
-          .bind("FunctionCallUsedNotInitializeVarDecl"),
-      this);
-
-  MF.addMatcher(
-      callExpr(
-          allOf(callee(functionDecl(functionName())), hasAncestor(varDecl()),
-                hasAncestor(functionDecl(unless(allOf(
-                    hasAttr(attr::CUDADevice), hasAttr(attr::CUDAGlobal)))))))
-          .bind("FunctionCallUsedInitializeVarDecl"),
-      this);
-
-  MF.addMatcher(
-      unresolvedLookupExpr(
-          hasAnyDeclaration(namedDecl(functionName())),
-          hasParent(callExpr(unless(parentStmt())).bind("callExprUsed")))
-          .bind("unresolvedCallUsed"),
-      this);
-}
-
-void BLASFunctionCallRule::runRule(const MatchFinder::MatchResult &Result) {
-  auto getArgWithTypeCast = [&](const Expr* E, const std::string& CastType) {
-    if (auto Cast = dyn_cast<CStyleCastExpr>(E->IgnoreImpCasts())) {
-      return "(" + CastType + ")" + ExprAnalysis::ref(Cast->getSubExpr());
-    } else {
-      return "(" + CastType + ")" + ExprAnalysis::ref(E);
-    }
-  };
-
-  bool IsAssigned = false;
-  bool IsInitializeVarDecl = false;
-  bool HasDeviceAttr = false;
-  std::string FuncName = "";
-  const CallExpr *CE = getNodeAsType<CallExpr>(Result, "kernelCall");
-  if (CE) {
-    HasDeviceAttr = true;
-  } else if (!(CE = getNodeAsType<CallExpr>(Result, "FunctionCall"))) {
-    if ((CE = getNodeAsType<CallExpr>(
-             Result, "FunctionCallUsedNotInitializeVarDecl"))) {
-      IsAssigned = true;
-    } else if ((CE = getNodeAsType<CallExpr>(
-                    Result, "FunctionCallUsedInitializeVarDecl"))) {
-      IsAssigned = true;
-      IsInitializeVarDecl = true;
-    } else if (auto *ULE = getNodeAsType<UnresolvedLookupExpr>(
-                   Result, "unresolvedCallUsed")) {
-      CE = getAssistNodeAsType<CallExpr>(Result, "callExprUsed");
-      FuncName = ULE->getName().getAsString();
-    } else {
-      return;
-    }
-  }
-
-  if (FuncName == "") {
-    if (!CE->getDirectCallee())
-      return;
-    FuncName = CE->getDirectCallee()->getNameInfo().getName().getAsString();
-  }
-
-  const SourceManager *SM = Result.SourceManager;
-  auto Loc = DpctGlobalInfo::getLocInfo(SM->getExpansionLoc(CE->getBeginLoc()));
-  DpctGlobalInfo::updateInitSuffixIndexInRule(
-      DpctGlobalInfo::getSuffixIndexInitValue(
-          Loc.first.getCanonicalPath().str() + std::to_string(Loc.second)));
-
-  SourceLocation FuncNameBegin(CE->getBeginLoc());
-  SourceLocation FuncCallEnd(CE->getEndLoc());
-  // There are some macros like "#define API API_v2"
-  // so the function names we match should have the
-  // suffix "_v2".
-  bool IsMacroArg = SM->isMacroArgExpansion(CE->getBeginLoc()) &&
-                    SM->isMacroArgExpansion(CE->getEndLoc());
-
-  if (FuncNameBegin.isMacroID() && IsMacroArg) {
-    FuncNameBegin = SM->getImmediateSpellingLoc(FuncNameBegin);
-    FuncNameBegin = SM->getExpansionLoc(FuncNameBegin);
-  } else if (FuncNameBegin.isMacroID()) {
-    FuncNameBegin = SM->getExpansionLoc(FuncNameBegin);
-  }
-
-  if (FuncCallEnd.isMacroID() && IsMacroArg) {
-    FuncCallEnd = SM->getImmediateSpellingLoc(FuncCallEnd);
-    FuncCallEnd = SM->getExpansionLoc(FuncCallEnd);
-  } else if (FuncCallEnd.isMacroID()) {
-    FuncCallEnd = SM->getExpansionLoc(FuncCallEnd);
-  }
-
-  // Offset 1 is the length of the last token ")"
-  FuncCallEnd = FuncCallEnd.getLocWithOffset(1);
-  auto SR = getScopeInsertRange(CE, FuncNameBegin, FuncCallEnd);
-  SourceLocation PrefixInsertLoc = SR.getBegin(), SuffixInsertLoc = SR.getEnd();
-
-  auto FuncCallLength =
-      SM->getCharacterData(FuncCallEnd) - SM->getCharacterData(FuncNameBegin);
-
-  bool CanAvoidUsingLambda = false;
-  SourceLocation OuterInsertLoc;
-  std::string OriginStmtType;
-  bool NeedUseLambda = isConditionOfFlowControl(
-      CE, OriginStmtType, CanAvoidUsingLambda, OuterInsertLoc);
-  bool IsInReturnStmt = isInReturnStmt(CE, OuterInsertLoc);
-  bool CanAvoidBrace = false;
-  const CompoundStmt *CS = findImmediateBlock(CE);
-  if (CS && (CS->size() == 1)) {
-    const Stmt *S = *(CS->child_begin());
-    if (CE == S || dyn_cast<ReturnStmt>(S))
-      CanAvoidBrace = true;
-  }
-
-  if (NeedUseLambda) {
-    PrefixInsertLoc = FuncNameBegin;
-    SuffixInsertLoc = FuncCallEnd;
-  } else if (IsMacroArg) {
-    NeedUseLambda = true;
-    SourceRange SR = getFunctionRange(CE);
-    PrefixInsertLoc = SR.getBegin();
-    SuffixInsertLoc = SR.getEnd();
-  } else if (IsInReturnStmt) {
-    NeedUseLambda = true;
-    CanAvoidUsingLambda = true;
-    OriginStmtType = "return";
-    // For some Legacy BLAS API (return the calculated value), below two
-    // variables are needed. Although the function call is in return stmt, it
-    // cannot move out and must use lambda.
-    PrefixInsertLoc = FuncNameBegin;
-    SuffixInsertLoc = FuncCallEnd;
-  }
-
-  if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_None &&
-      (FuncName == "cublasHgemmBatched" || FuncName == "cublasSgemmBatched" ||
-       FuncName == "cublasDgemmBatched" || FuncName == "cublasCgemmBatched" ||
-       FuncName == "cublasZgemmBatched" || FuncName == "cublasStrsmBatched" ||
-       FuncName == "cublasDtrsmBatched" || FuncName == "cublasCtrsmBatched" ||
-       FuncName == "cublasZtrsmBatched" || FuncName == "cublasGemmBatchedEx" ||
-       FuncName == "cublasGemmBatchedEx_64")) {
-    report(FuncNameBegin, Diagnostics::API_NOT_MIGRATED, false, FuncName);
-    return;
-  }
-
-  std::string IndentStr = getIndent(PrefixInsertLoc, *SM).str();
-  // PrefixInsertStr: stmt + NL + indent
-  // SuffixInsertStr: NL + indent + stmt
-  std::string PrefixInsertStr, SuffixInsertStr;
-  // Clean it before starting migration
-  CallExprReplStr = "";
-  // TODO: Need to process the situation when scalar pointers (alpha, beta)
-  // are device pointers.
-
-  auto Item = MapNames::BLASAPIWithRewriter.find(FuncName);
-  if (Item != MapNames::BLASAPIWithRewriter.end()) {
-    std::string NewFunctionName = Item->second;
-    if (HasDeviceAttr && !NewFunctionName.empty()) {
-      report(FuncNameBegin, Diagnostics::FUNCTION_CALL_IN_DEVICE, false,
-             MapNames::ITFName.at(FuncName), NewFunctionName);
-      return;
-    }
-    ExprAnalysis EA(CE);
-    emplaceTransformation(EA.getReplacement());
-    EA.applyAllSubExprRepl();
-    return;
-  } else if (MapNames::LegacyBLASFuncReplInfoMap.find(FuncName) !=
-             MapNames::LegacyBLASFuncReplInfoMap.end()) {
-    auto ReplInfoPair = MapNames::LegacyBLASFuncReplInfoMap.find(FuncName);
-    MapNames::BLASFuncComplexReplInfo ReplInfo = ReplInfoPair->second;
-    requestFeature(HelperFeatureEnum::device_ext);
-    CallExprReplStr = CallExprReplStr + ReplInfo.ReplName + "(" +
-                      MapNames::getLibraryHelperNamespace() +
-                      "blas::descriptor::get_saved_queue()";
-    std::string IndentStr =
-        getIndent(PrefixInsertLoc, (Result.Context)->getSourceManager()).str();
-
-    std::string VarType;
-    std::string VarName;
-    std::string DeclOutOfBrace;
-    const VarDecl *VD = 0;
-    if (IsInitializeVarDecl) {
-      VD = getAncestralVarDecl(CE);
-      if (VD) {
-        VarType = VD->getType().getAsString();
-        if (VarType == "cuComplex" || VarType == "cuFloatComplex") {
-          VarType = MapNames::getClNamespace() + "float2";
-        }
-        if (VarType == "cuDoubleComplex") {
-          VarType = MapNames::getClNamespace() + "double2";
-        }
-        VarName = VD->getNameAsString();
-      } else {
-        assert(0 && "Fail to get VarDecl information");
-        return;
-      }
-      DeclOutOfBrace = VarType + " " + VarName + ";" + getNL() + IndentStr;
-    }
-    std::vector<std::string> ParamsStrsVec =
-        getParamsAsStrs(CE, *(Result.Context));
-    int ArgNum = CE->getNumArgs();
-    for (int i = 0; i < ArgNum; ++i) {
-      int IndexTemp = -1;
-      if (isReplIndex(i, ReplInfo.BufferIndexInfo, IndexTemp)) {
-        if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_Restricted) {
-          if ((FuncName == "cublasSrotm" || FuncName == "cublasDrotm") &&
-              i == 5) {
-            CallExprReplStr = CallExprReplStr + ", const_cast<" +
-                              ReplInfo.BufferTypeInfo[IndexTemp] + "*>(" +
-                              ExprAnalysis::ref(CE->getArg(5)) + ")";
-          } else if (ReplInfo.BufferTypeInfo[IndexTemp] ==
-                         "std::complex<float>" ||
-                     ReplInfo.BufferTypeInfo[IndexTemp] ==
-                         "std::complex<double>") {
-            CallExprReplStr =
-                CallExprReplStr + ", " +
-                getArgWithTypeCast(CE->getArg(i),
-                                   ReplInfo.BufferTypeInfo[IndexTemp] + "*");
-          } else {
-            CallExprReplStr = CallExprReplStr + ", " + ParamsStrsVec[i];
-          }
-        } else {
-          requestFeature(HelperFeatureEnum::device_ext);
-          std::string BufferDecl;
-          std::string BufferName = getBufferNameAndDeclStr(
-              CE->getArg(i), ReplInfo.BufferTypeInfo[IndexTemp], IndentStr,
-              BufferDecl);
-          CallExprReplStr = CallExprReplStr + ", " + BufferName;
-          PrefixInsertStr = PrefixInsertStr + BufferDecl;
-        }
-      } else if (isReplIndex(i, ReplInfo.PointerIndexInfo, IndexTemp)) {
-        if (ReplInfo.PointerTypeInfo[IndexTemp] == "float" ||
-            ReplInfo.PointerTypeInfo[IndexTemp] == "double") {
-          // This code path is only for legacy cublasSrotmg and cublasDrotmg
-          CallExprReplStr = CallExprReplStr + ", " +
-                            getDrefName(CE->getArg(i)->IgnoreImplicit());
-        } else {
-          if (isAnIdentifierOrLiteral(CE->getArg(i)))
-            CallExprReplStr =
-                CallExprReplStr + ", " + ReplInfo.PointerTypeInfo[IndexTemp] +
-                "(" + ParamsStrsVec[i] + ".x()," + ParamsStrsVec[i] + ".y())";
-          else
-            CallExprReplStr = CallExprReplStr + ", " +
-                              ReplInfo.PointerTypeInfo[IndexTemp] + "((" +
-                              ParamsStrsVec[i] + ").x(),(" + ParamsStrsVec[i] +
-                              ").y())";
-        }
-      } else if (isReplIndex(i, ReplInfo.OperationIndexInfo, IndexTemp)) {
-        Expr::EvalResult ER;
-        if (CE->getArg(i)->EvaluateAsInt(ER, *Result.Context) &&
-            !CE->getArg(i)->getBeginLoc().isMacroID()) {
-          int64_t Value = ER.Val.getInt().getExtValue();
-          if (Value == 'N' || Value == 'n') {
-            CallExprReplStr =
-                CallExprReplStr + ", oneapi::mkl::transpose::nontrans";
-          } else if (Value == 'T' || Value == 't') {
-            CallExprReplStr =
-                CallExprReplStr + ", oneapi::mkl::transpose::trans";
-          } else {
-            CallExprReplStr =
-                CallExprReplStr + ", oneapi::mkl::transpose::conjtrans";
-          }
-        } else {
-          std::string TransParamName;
-          if (CE->getArg(i)->HasSideEffects(DpctGlobalInfo::getContext())) {
-            TransParamName =
-                "transpose_ct" +
-                std::to_string(DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-            PrefixInsertStr = PrefixInsertStr + "auto " + TransParamName +
-                              " = " + ParamsStrsVec[i] + ";" + getNL() +
-                              IndentStr;
-          } else {
-            TransParamName = ParamsStrsVec[i];
-          }
-          CallExprReplStr = CallExprReplStr + ", " + "(" + TransParamName +
-                            "=='N'||" + TransParamName +
-                            "=='n') ? oneapi::mkl::transpose::nontrans: ((" +
-                            TransParamName + "=='T'||" + TransParamName +
-                            "=='t') ? oneapi::mkl::transpose::"
-                            "trans : oneapi::mkl::transpose::conjtrans)";
-        }
-      } else if (ReplInfo.FillModeIndexInfo == i) {
-        Expr::EvalResult ER;
-        if (CE->getArg(i)->EvaluateAsInt(ER, *Result.Context) &&
-            !CE->getArg(i)->getBeginLoc().isMacroID()) {
-          int64_t Value = ER.Val.getInt().getExtValue();
-          if (Value == 'U' || Value == 'u') {
-            CallExprReplStr = CallExprReplStr + ", oneapi::mkl::uplo::upper";
-          } else {
-            CallExprReplStr = CallExprReplStr + ", oneapi::mkl::uplo::lower";
-          }
-        } else {
-          std::string FillParamName;
-          if (CE->getArg(i)->HasSideEffects(DpctGlobalInfo::getContext())) {
-            FillParamName =
-                "fillmode_ct" +
-                std::to_string(DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-            PrefixInsertStr = PrefixInsertStr + "auto " + FillParamName +
-                              " = " + ParamsStrsVec[i] + ";" + getNL() +
-                              IndentStr;
-          } else {
-            FillParamName = ParamsStrsVec[i];
-          }
-          CallExprReplStr =
-              CallExprReplStr + ", " + "(" + FillParamName + "=='L'||" +
-              FillParamName +
-              "=='l') ? oneapi::mkl::uplo::lower : oneapi::mkl::uplo::upper";
-        }
-      } else if (ReplInfo.SideModeIndexInfo == i) {
-        Expr::EvalResult ER;
-        if (CE->getArg(i)->EvaluateAsInt(ER, *Result.Context) &&
-            !CE->getArg(i)->getBeginLoc().isMacroID()) {
-          int64_t Value = ER.Val.getInt().getExtValue();
-          if (Value == 'L' || Value == 'l') {
-            CallExprReplStr = CallExprReplStr + ", oneapi::mkl::side::left";
-          } else {
-            CallExprReplStr = CallExprReplStr + ", oneapi::mkl::side::right";
-          }
-        } else {
-          std::string SideParamName;
-          if (CE->getArg(i)->HasSideEffects(DpctGlobalInfo::getContext())) {
-            SideParamName =
-                "sidemode_ct" +
-                std::to_string(DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-            PrefixInsertStr = PrefixInsertStr + "auto " + SideParamName +
-                              " = " + ParamsStrsVec[i] + ";" + getNL() +
-                              IndentStr;
-          } else {
-            SideParamName = ParamsStrsVec[i];
-          }
-          CallExprReplStr =
-              CallExprReplStr + ", " + "(" + SideParamName + "=='L'||" +
-              SideParamName +
-              "=='l') ? oneapi::mkl::side::left : oneapi::mkl::side::right";
-        }
-      } else if (ReplInfo.DiagTypeIndexInfo == i) {
-        Expr::EvalResult ER;
-        if (CE->getArg(i)->EvaluateAsInt(ER, *Result.Context) &&
-            !CE->getArg(i)->getBeginLoc().isMacroID()) {
-          int64_t Value = ER.Val.getInt().getExtValue();
-          if (Value == 'N' || Value == 'n') {
-            CallExprReplStr = CallExprReplStr + ", oneapi::mkl::diag::nonunit";
-          } else {
-            CallExprReplStr = CallExprReplStr + ", oneapi::mkl::diag::unit";
-          }
-        } else {
-          std::string DiagParamName;
-          if (CE->getArg(i)->HasSideEffects(DpctGlobalInfo::getContext())) {
-            DiagParamName =
-                "diagtype_ct" +
-                std::to_string(DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-            PrefixInsertStr = PrefixInsertStr + "auto " + DiagParamName +
-                              " = " + ParamsStrsVec[i] + ";" + getNL() +
-                              IndentStr;
-          } else {
-            DiagParamName = ParamsStrsVec[i];
-          }
-          CallExprReplStr =
-              CallExprReplStr + ", " + "(" + DiagParamName + "=='N'||" +
-              DiagParamName +
-              "=='n') ? oneapi::mkl::diag::nonunit : oneapi::mkl::diag::unit";
-        }
-      } else {
-        CallExprReplStr = CallExprReplStr + ", " + ParamsStrsVec[i];
-      }
-    }
-
-    // All legacy APIs are synchronous
-    if (FuncName == "cublasSnrm2" || FuncName == "cublasDnrm2" ||
-        FuncName == "cublasScnrm2" || FuncName == "cublasDznrm2" ||
-        FuncName == "cublasSdot" || FuncName == "cublasDdot" ||
-        FuncName == "cublasCdotu" || FuncName == "cublasCdotc" ||
-        FuncName == "cublasZdotu" || FuncName == "cublasZdotc" ||
-        FuncName == "cublasIsamax" || FuncName == "cublasIdamax" ||
-        FuncName == "cublasIcamax" || FuncName == "cublasIzamax" ||
-        FuncName == "cublasIsamin" || FuncName == "cublasIdamin" ||
-        FuncName == "cublasIcamin" || FuncName == "cublasIzamin" ||
-        FuncName == "cublasSasum" || FuncName == "cublasDasum" ||
-        FuncName == "cublasScasum" || FuncName == "cublasDzasum") {
-      // APIs which have return value
-      std::string ResultTempPtr =
-          "res_temp_ptr_ct" +
-          std::to_string(DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-      std::string ResultTempVal =
-          "res_temp_val_ct" +
-          std::to_string(DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-      std::string ResultTempBuf =
-          "res_temp_buf_ct" +
-          std::to_string(DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-      std::string ResultType =
-          ReplInfo.BufferTypeInfo[ReplInfo.BufferTypeInfo.size() - 1];
-      std::string ReturnValueParamsStr;
-      if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_Restricted) {
-        requestFeature(HelperFeatureEnum::device_ext);
-        auto DefaultQueue = DpctGlobalInfo::getDefaultQueue(CE);
-        PrefixInsertStr = PrefixInsertStr + ResultType + "* " + ResultTempPtr +
-                          " = " + MapNames::getClNamespace() +
-                          "malloc_shared<" + ResultType + ">(1, " + DefaultQueue + ");" +
-                          getNL() + IndentStr + CallExprReplStr + ", " +
-                          ResultTempPtr + ").wait();" + getNL() + IndentStr;
-
-        ReturnValueParamsStr =
-            "(" + ResultTempPtr + "->real(), " + ResultTempPtr + "->imag())";
-
-        if (NeedUseLambda) {
-          PrefixInsertStr = PrefixInsertStr + ResultType + " " + ResultTempVal +
-                            " = *" + ResultTempPtr + ";" + getNL() + IndentStr +
-                            MapNames::getClNamespace() + "free(" +
-                            ResultTempPtr + ", " + DefaultQueue + ");" +
-                            getNL() + IndentStr;
-          ReturnValueParamsStr =
-              "(" + ResultTempVal + ".real(), " + ResultTempVal + ".imag())";
-        } else {
-          SuffixInsertStr = SuffixInsertStr + getNL() + IndentStr +
-                            MapNames::getClNamespace() + "free(" +
-                            ResultTempPtr + ", " + DefaultQueue + ");";
-        }
-      } else {
-        PrefixInsertStr = PrefixInsertStr + MapNames::getClNamespace() +
-                          "buffer<" + ResultType + "> " + ResultTempBuf + "(" +
-                          MapNames::getClNamespace() + "range<1>(1));" +
-                          getNL() + IndentStr + CallExprReplStr + ", " +
-                          ResultTempBuf + ");" + getNL() + IndentStr;
-        ReturnValueParamsStr =
-            "(" + ResultTempBuf + ".get_access<" + MapNames::getClNamespace() +
-            "access_mode::read>()[0].real(), " + ResultTempBuf +
-            ".get_access<" + MapNames::getClNamespace() +
-            "access_mode::read>()[0].imag())";
-      }
-
-      std::string Repl;
-      if (FuncName == "cublasCdotu" || FuncName == "cublasCdotc") {
-        Repl = MapNames::getClNamespace() + "float2" + ReturnValueParamsStr;
-      } else if (FuncName == "cublasZdotu" || FuncName == "cublasZdotc") {
-        Repl = MapNames::getClNamespace() + "double2" + ReturnValueParamsStr;
-      } else {
-        if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_Restricted) {
-          if (NeedUseLambda)
-            Repl = ResultTempVal;
-          else
-            Repl = "*" + ResultTempPtr;
-        } else {
-          Repl = ResultTempBuf + ".get_access<" + MapNames::getClNamespace() +
-                 "access_mode::read>()[0]";
-        }
-      }
-      if (NeedUseLambda) {
-        std::string CallRepl = "return " + Repl;
-
-        insertAroundRange(PrefixInsertLoc, SuffixInsertLoc,
-                          std::string("[&](){") + getNL() + IndentStr +
-                              PrefixInsertStr,
-                          ";" + SuffixInsertStr + getNL() + IndentStr + "}()");
-        emplaceTransformation(new ReplaceStmt(CE, CallRepl));
-      } else {
-        if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_None)
-          insertAroundRange(PrefixInsertLoc, SuffixInsertLoc,
-                            DeclOutOfBrace + "{" + getNL() + IndentStr +
-                                PrefixInsertStr,
-                            SuffixInsertStr + getNL() + IndentStr + "}");
-        else
-          insertAroundRange(PrefixInsertLoc, SuffixInsertLoc,
-                            DeclOutOfBrace + PrefixInsertStr,
-                            std::move(SuffixInsertStr));
-
-        if (IsInitializeVarDecl) {
-          auto ParentNodes = (Result.Context)->getParents(*VD);
-          const DeclStmt *DS = 0;
-          if ((DS = ParentNodes[0].get<DeclStmt>())) {
-            emplaceTransformation(
-                new ReplaceStmt(DS, VarName + " = " + Repl + ";"));
-          } else {
-            assert(0 && "Fail to get Var Decl Stmt");
-            return;
-          }
-        } else {
-          emplaceTransformation(new ReplaceStmt(CE, Repl));
-        }
-      }
-    } else {
-      // APIs which haven't return value
-      if (NeedUseLambda) {
-        if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_Restricted) {
-          CallExprReplStr = CallExprReplStr + ").wait()";
-        } else {
-          CallExprReplStr = CallExprReplStr + ")";
-        }
-        if (CanAvoidUsingLambda) {
-          std::string InsertStr;
-          if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_None)
-            InsertStr = DeclOutOfBrace + "{" + getNL() + IndentStr +
-                        PrefixInsertStr + CallExprReplStr + ";" +
-                        SuffixInsertStr + getNL() + IndentStr + "}" + getNL() +
-                        IndentStr;
-          else
-            InsertStr = DeclOutOfBrace + PrefixInsertStr + CallExprReplStr +
-                        ";" + SuffixInsertStr + getNL() + IndentStr;
-          emplaceTransformation(
-              new InsertText(OuterInsertLoc, std::move(InsertStr)));
-          // APIs in this code path haven't return value, so remove the CallExpr
-          emplaceTransformation(
-              new ReplaceText(FuncNameBegin, FuncCallLength, ""));
-        } else {
-          emplaceTransformation(
-              new ReplaceStmt(CE, std::move(CallExprReplStr)));
-          insertAroundRange(PrefixInsertLoc, SuffixInsertLoc,
-                            std::string("[&](){") + getNL() + IndentStr +
-                                PrefixInsertStr,
-                            getNL() + IndentStr + std::string("}()"));
-        }
-      } else {
-        if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_Restricted) {
-          CallExprReplStr = CallExprReplStr + ").wait()";
-        } else {
-          CallExprReplStr = CallExprReplStr + ")";
-        }
-        emplaceTransformation(new ReplaceStmt(CE, std::move(CallExprReplStr)));
-        if (!PrefixInsertStr.empty()) {
-          if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_None)
-            insertAroundRange(PrefixInsertLoc, SuffixInsertLoc,
-                              std::string("{") + getNL() + IndentStr +
-                                  PrefixInsertStr,
-                              getNL() + IndentStr + std::string("}"));
-          else
-            insertAroundRange(PrefixInsertLoc, SuffixInsertLoc,
-                              std::move(PrefixInsertStr), "");
-        }
-      }
-    }
-  } else if (FuncName == "cublasInit" || FuncName == "cublasShutdown" ||
-             FuncName == "cublasGetError") {
-    // Remove these three function calls.
-    // TODO: migrate functions when they are in template
-    auto Msg = MapNames::RemovedAPIWarningMessage.find(FuncName);
-    SourceRange SR = getFunctionRange(CE);
-    auto Len = SM->getDecomposedLoc(SR.getEnd()).second -
-               SM->getDecomposedLoc(SR.getBegin()).second;
-    if (SM->isMacroArgExpansion(CE->getBeginLoc()) &&
-        SM->isMacroArgExpansion(CE->getEndLoc())) {
-      if (IsAssigned) {
-        report(CE->getBeginLoc(), Diagnostics::FUNC_CALL_REMOVED_0, false,
-               MapNames::ITFName.at(FuncName), Msg->second);
-        emplaceTransformation(new ReplaceText(SR.getBegin(), Len, "0"));
-      } else {
-        report(CE->getBeginLoc(), Diagnostics::FUNC_CALL_REMOVED, false,
-               MapNames::ITFName.at(FuncName), Msg->second);
-        emplaceTransformation(new ReplaceText(SR.getBegin(), Len, "0"));
-      }
-    } else {
-      if (IsAssigned) {
-        report(CE->getBeginLoc(), Diagnostics::FUNC_CALL_REMOVED_0, false,
-               MapNames::ITFName.at(FuncName), Msg->second);
-        emplaceTransformation(new ReplaceStmt(CE, false, "0"));
-      } else {
-        report(CE->getBeginLoc(), Diagnostics::FUNC_CALL_REMOVED, false,
-               MapNames::ITFName.at(FuncName), Msg->second);
-        emplaceTransformation(new ReplaceStmt(CE, false, ""));
-      }
-    }
-  } else if (FuncName == "cublasGetPointerMode_v2" ||
-             FuncName == "cublasSetPointerMode_v2" ||
-             FuncName == "cublasGetAtomicsMode" ||
-             FuncName == "cublasSetAtomicsMode" ||
-             FuncName == "cublasSetWorkspace_v2") {
-    std::string Msg = "this functionality is redundant in SYCL.";
-    if (IsAssigned) {
-      report(CE->getBeginLoc(), Diagnostics::FUNC_CALL_REMOVED_0, false,
-             MapNames::ITFName.at(FuncName), Msg);
-      emplaceTransformation(new ReplaceStmt(CE, true, "0"));
-    } else {
-      report(CE->getBeginLoc(), Diagnostics::FUNC_CALL_REMOVED, false,
-             MapNames::ITFName.at(FuncName), Msg);
-      emplaceTransformation(new ReplaceStmt(CE, true, ""));
-    }
-  } else if (FuncName == "cublasSetVector" || FuncName == "cublasGetVector" ||
-             FuncName == "cublasSetVectorAsync" ||
-             FuncName == "cublasGetVectorAsync" ||
-             FuncName == "cublasSetVector_64" ||
-             FuncName == "cublasGetVector_64" ||
-             FuncName == "cublasSetVectorAsync_64" ||
-             FuncName == "cublasGetVectorAsync_64") {
-    if (HasDeviceAttr) {
-      report(CE->getBeginLoc(), Diagnostics::FUNCTION_CALL_IN_DEVICE, false,
-             MapNames::ITFName.at(FuncName),
-             MapNames::getLibraryHelperNamespace() + "blas::matrix_mem_copy");
-      return;
-    }
-    const Expr *IncxExpr = CE->getArg(3);
-    const Expr *IncyExpr = CE->getArg(5);
-    const std::string MigratedIncx = ExprAnalysis::ref(IncxExpr);
-    const std::string MigratedIncy = ExprAnalysis::ref(IncyExpr);
-    Expr::EvalResult IncxExprResult, IncyExprResult;
-    if (IncxExpr->EvaluateAsInt(IncxExprResult, *Result.Context) &&
-        IncyExpr->EvaluateAsInt(IncyExprResult, *Result.Context)) {
-      int64_t IncxValue = IncxExprResult.Val.getInt().getExtValue();
-      int64_t IncyValue = IncyExprResult.Val.getInt().getExtValue();
-      if (IncxValue != IncyValue) {
-        report(CE->getBeginLoc(), Diagnostics::POTENTIAL_PERFORMANCE_ISSUE,
-               false, MapNames::ITFName.at(FuncName),
-               "parameter " + MigratedIncx + " does not equal to parameter " +
-                   MigratedIncy);
-      } else if ((IncxValue == IncyValue) && (IncxValue != 1)) {
-        // incx equals to incy, but does not equal to 1. Performance issue may
-        // occur.
-        report(CE->getBeginLoc(), Diagnostics::POTENTIAL_PERFORMANCE_ISSUE,
-               false, MapNames::ITFName.at(FuncName),
-               "parameter " + MigratedIncx + " equals to parameter " +
-                   MigratedIncy + " but greater than 1");
-      }
-    } else {
-      report(CE->getBeginLoc(), Diagnostics::POTENTIAL_PERFORMANCE_ISSUE, false,
-             MapNames::ITFName.at(FuncName),
-             "parameter(s) " + MigratedIncx + " and/or " + MigratedIncy +
-                 " could not be evaluated");
-    }
-
-    ExprAnalysis EA(CE);
-    emplaceTransformation(EA.getReplacement());
-    EA.applyAllSubExprRepl();
-    return;
-  } else if (FuncName == "cublasSetMatrix" || FuncName == "cublasGetMatrix" ||
-             FuncName == "cublasSetMatrixAsync" ||
-             FuncName == "cublasGetMatrixAsync" ||
-             FuncName == "cublasSetMatrix_64" ||
-             FuncName == "cublasGetMatrix_64" ||
-             FuncName == "cublasSetMatrixAsync_64" ||
-             FuncName == "cublasGetMatrixAsync_64") {
-    if (HasDeviceAttr) {
-      report(CE->getBeginLoc(), Diagnostics::FUNCTION_CALL_IN_DEVICE, false,
-             MapNames::ITFName.at(FuncName),
-             MapNames::getLibraryHelperNamespace() + "blas::matrix_mem_copy");
-      return;
-    }
-    const Expr *RowsExpr = CE->getArg(0);
-    const Expr *LdaExpr = CE->getArg(4);
-    const Expr *LdbExpr = CE->getArg(6);
-    const std::string MigratedRows = ExprAnalysis::ref(RowsExpr);
-    const std::string MigratedLda = ExprAnalysis::ref(LdaExpr);
-    const std::string MigratedLdb = ExprAnalysis::ref(LdbExpr);
-    Expr::EvalResult LdaExprResult, LdbExprResult;
-    if (LdaExpr->EvaluateAsInt(LdaExprResult, *Result.Context) &&
-        LdbExpr->EvaluateAsInt(LdbExprResult, *Result.Context)) {
-      int64_t LdaValue = LdaExprResult.Val.getInt().getExtValue();
-      int64_t LdbValue = LdbExprResult.Val.getInt().getExtValue();
-      if (LdaValue != LdbValue) {
-        report(CE->getBeginLoc(), Diagnostics::POTENTIAL_PERFORMANCE_ISSUE,
-               false, MapNames::ITFName.at(FuncName),
-               "parameter " + MigratedLda + " does not equal to parameter " +
-                   MigratedLdb);
-      } else {
-        Expr::EvalResult RowsExprResult;
-        if (RowsExpr->EvaluateAsInt(RowsExprResult, *Result.Context)) {
-          int64_t RowsValue = RowsExprResult.Val.getInt().getExtValue();
-          if (LdaValue > RowsValue) {
-            // lda > rows. Performance issue may occur.
-            report(CE->getBeginLoc(), Diagnostics::POTENTIAL_PERFORMANCE_ISSUE,
-                   false, MapNames::ITFName.at(FuncName),
-                   "parameter " + MigratedRows + " is smaller than parameter " +
-                       MigratedLda);
-          }
-        } else {
-          // rows cannot be evaluated. Performance issue may occur.
-          report(
-              CE->getBeginLoc(), Diagnostics::POTENTIAL_PERFORMANCE_ISSUE,
-              false, MapNames::ITFName.at(FuncName),
-              "parameter " + MigratedRows +
-                  " could not be evaluated and may be smaller than parameter " +
-                  MigratedLda);
-        }
-      }
-    } else {
-      report(CE->getBeginLoc(), Diagnostics::POTENTIAL_PERFORMANCE_ISSUE, false,
-             MapNames::ITFName.at(FuncName),
-             "parameter(s) " + MigratedLda + " and/or " + MigratedLdb +
-                 " could not be evaluated");
-    }
-
-    ExprAnalysis EA(CE);
-    emplaceTransformation(EA.getReplacement());
-    EA.applyAllSubExprRepl();
-    return;
-  } else if (FuncName == "cublasGetVersion" ||
-             FuncName == "cublasGetVersion_v2") {
-    if (FuncName == "cublasGetVersion" || FuncName == "cublasGetVersion_v2") {
-      DpctGlobalInfo::getInstance().insertHeader(
-          SM->getExpansionLoc(CE->getBeginLoc()), HT_DPCT_COMMON_Utils);
-    }
-
-    ExprAnalysis EA(CE);
-    emplaceTransformation(EA.getReplacement());
-    EA.applyAllSubExprRepl();
-  } else {
-    assert(0 && "Unknown function name");
-  }
-}
-
-// Check whether the input expression is CallExpr or UnaryExprOrTypeTraitExpr
-// (sizeof or alignof) or an identifier or literal
-bool BLASFunctionCallRule::isCEOrUETTEOrAnIdentifierOrLiteral(const Expr *E) {
-  auto CE = dyn_cast<CallExpr>(E->IgnoreImpCasts());
-  if (CE != nullptr) {
-    return true;
-  }
-  auto UETTE = dyn_cast<UnaryExprOrTypeTraitExpr>(E->IgnoreImpCasts());
-  if (UETTE != nullptr) {
-    return true;
-  }
-  if (isAnIdentifierOrLiteral(E)) {
-    return true;
-  }
-  return false;
-}
-
-bool BLASFunctionCallRule::isReplIndex(int Input,
-                                       const std::vector<int> &IndexInfo,
-                                       int &IndexTemp) {
-  for (int i = 0; i < static_cast<int>(IndexInfo.size()); ++i) {
-    if (IndexInfo[i] == Input) {
-      IndexTemp = i;
-      return true;
-    }
-  }
-  return false;
-}
-
-std::vector<std::string>
-BLASFunctionCallRule::getParamsAsStrs(const CallExpr *CE,
-                                      const ASTContext &Context) {
-  std::vector<std::string> ParamsStrVec;
-  for (auto Arg : CE->arguments())
-    ParamsStrVec.emplace_back(ExprAnalysis::ref(Arg));
-  return ParamsStrVec;
-}
-
-// sample code looks like:
-//   Complex-type res1 = API(...);
-//   res2 = API(...);
-//
-// migrated code looks like:
-//   Complex-type res1;
-//   {
-//   buffer res_buffer;
-//   mklAPI(res_buffer);
-//   res1 = res_buffer.get_access()[0];
-//   }
-//   {
-//   buffer res_buffer;
-//   mklAPI(res_buffer);
-//   res2 = res_buffer.get_access()[0];
-//   }
-//
-// If the API return value initializes the var declaration, we need to put the
-// var declaration out of the scope and assign it in the scope, otherwise users
-// cannot use this var out of the scope.
-// So we need to find the Decl node of the var, which is the CallExpr node's
-// ancestor.
-// The initial node is the matched CallExpr node. Then visit the parent node of
-// the current node until the current node is a VarDecl node.
-const clang::VarDecl *
-BLASFunctionCallRule::getAncestralVarDecl(const clang::CallExpr *CE) {
-  auto &Context = dpct::DpctGlobalInfo::getContext();
-  auto Parents = Context.getParents(*CE);
-  while (Parents.size() == 1) {
-    auto *Parent = Parents[0].get<VarDecl>();
-    if (Parent) {
-      return Parent;
-    } else {
-      Parents = Context.getParents(Parents[0]);
-    }
-  }
-  return nullptr;
-}
-
-// Rule for SOLVER enums.
-// Migrate SOLVER status values to corresponding int values
-// Other SOLVER named values are migrated to corresponding named values
-void SOLVEREnumsRule::registerMatcher(MatchFinder &MF) {
-  MF.addMatcher(declRefExpr(to(enumConstantDecl(matchesName(
-                                "(CUSOLVER_STATU.*)|(CUSOLVER_ALG.*)|("
-                                "CUSOLVERDN_GETRF)|(CUSOLVERDN_POTRF)"))))
-                    .bind("SOLVERConstants"),
-                this);
-  MF.addMatcher(
-      declRefExpr(to(enumConstantDecl(matchesName(
-                      "(CUSOLVER_EIG_TYPE.*)|(CUSOLVER_EIG_MODE.*)"))))
-          .bind("SLOVERNamedValueConstants"),
-      this);
-}
-
-void SOLVEREnumsRule::runRule(const MatchFinder::MatchResult &Result) {
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "SOLVERConstants")) {
-    auto *EC = cast<EnumConstantDecl>(DE->getDecl());
-    emplaceTransformation(new ReplaceStmt(DE, toString(EC->getInitVal(), 10)));
-  }
-
-  if (const DeclRefExpr *DE =
-          getNodeAsType<DeclRefExpr>(Result, "SLOVERNamedValueConstants")) {
-    auto *EC = cast<EnumConstantDecl>(DE->getDecl());
-    std::string Name = EC->getNameAsString();
-    auto Search = MapNames::SOLVEREnumsMap.find(Name);
-    if (Search == MapNames::SOLVEREnumsMap.end()) {
-      llvm::dbgs() << "[" << getName()
-                   << "] Unexpected enum variable: " << Name;
-      return;
-    }
-    std::string Replacement = Search->second;
-    emplaceTransformation(new ReplaceStmt(DE, std::move(Replacement)));
-  }
-}
-
-
-void SOLVERFunctionCallRule::registerMatcher(MatchFinder &MF) {
-  auto functionName = [&]() {
-    return hasAnyName(
-        "cusolverDnSetAdvOptions", "cusolverDnGetStream", "cusolverDnSetStream",
-        "cusolverDnCreateParams", "cusolverDnDestroyParams", "cusolverDnCreate",
-        "cusolverDnDestroy", "cusolverDnSpotrf_bufferSize",
-        "cusolverDnDpotrf_bufferSize", "cusolverDnCpotrf_bufferSize",
-        "cusolverDnZpotrf_bufferSize", "cusolverDnSpotri_bufferSize",
-        "cusolverDnDpotri_bufferSize", "cusolverDnCpotri_bufferSize",
-        "cusolverDnZpotri_bufferSize", "cusolverDnSgetrf_bufferSize",
-        "cusolverDnDgetrf_bufferSize", "cusolverDnCgetrf_bufferSize",
-        "cusolverDnZgetrf_bufferSize", "cusolverDnSpotrf", "cusolverDnDpotrf",
-        "cusolverDnCpotrf", "cusolverDnZpotrf", "cusolverDnSpotrs",
-        "cusolverDnDpotrs", "cusolverDnCpotrs", "cusolverDnZpotrs",
-        "cusolverDnSpotri", "cusolverDnDpotri", "cusolverDnCpotri",
-        "cusolverDnZpotri", "cusolverDnSgetrf", "cusolverDnDgetrf",
-        "cusolverDnCgetrf", "cusolverDnZgetrf", "cusolverDnSgetrs",
-        "cusolverDnDgetrs", "cusolverDnCgetrs", "cusolverDnZgetrs",
-        "cusolverDnSgeqrf_bufferSize", "cusolverDnDgeqrf_bufferSize",
-        "cusolverDnCgeqrf_bufferSize", "cusolverDnZgeqrf_bufferSize",
-        "cusolverDnSgeqrf", "cusolverDnDgeqrf", "cusolverDnCgeqrf",
-        "cusolverDnZgeqrf", "cusolverDnSormqr_bufferSize",
-        "cusolverDnDormqr_bufferSize", "cusolverDnSormqr", "cusolverDnDormqr",
-        "cusolverDnCunmqr_bufferSize", "cusolverDnZunmqr_bufferSize",
-        "cusolverDnCunmqr", "cusolverDnZunmqr", "cusolverDnSorgqr_bufferSize",
-        "cusolverDnDorgqr_bufferSize", "cusolverDnCungqr_bufferSize",
-        "cusolverDnZungqr_bufferSize", "cusolverDnSorgqr", "cusolverDnDorgqr",
-        "cusolverDnCungqr", "cusolverDnZungqr", "cusolverDnSsytrf_bufferSize",
-        "cusolverDnDsytrf_bufferSize", "cusolverDnCsytrf_bufferSize",
-        "cusolverDnZsytrf_bufferSize", "cusolverDnSsytrf", "cusolverDnDsytrf",
-        "cusolverDnCsytrf", "cusolverDnZsytrf", "cusolverDnSgebrd_bufferSize",
-        "cusolverDnDgebrd_bufferSize", "cusolverDnCgebrd_bufferSize",
-        "cusolverDnZgebrd_bufferSize", "cusolverDnSgebrd", "cusolverDnDgebrd",
-        "cusolverDnCgebrd", "cusolverDnZgebrd", "cusolverDnSorgbr_bufferSize",
-        "cusolverDnDorgbr_bufferSize", "cusolverDnCungbr_bufferSize",
-        "cusolverDnZungbr_bufferSize", "cusolverDnSorgbr", "cusolverDnDorgbr",
-        "cusolverDnCungbr", "cusolverDnZungbr", "cusolverDnSsytrd_bufferSize",
-        "cusolverDnDsytrd_bufferSize", "cusolverDnChetrd_bufferSize",
-        "cusolverDnZhetrd_bufferSize", "cusolverDnSsytrd", "cusolverDnDsytrd",
-        "cusolverDnChetrd", "cusolverDnZhetrd", "cusolverDnSormtr_bufferSize",
-        "cusolverDnDormtr_bufferSize", "cusolverDnCunmtr_bufferSize",
-        "cusolverDnZunmtr_bufferSize", "cusolverDnSormtr", "cusolverDnDormtr",
-        "cusolverDnCunmtr", "cusolverDnZunmtr", "cusolverDnSorgtr_bufferSize",
-        "cusolverDnDorgtr_bufferSize", "cusolverDnCungtr_bufferSize",
-        "cusolverDnZungtr_bufferSize", "cusolverDnSorgtr", "cusolverDnDorgtr",
-        "cusolverDnCungtr", "cusolverDnZungtr", "cusolverDnSgesvd_bufferSize",
-        "cusolverDnDgesvd_bufferSize", "cusolverDnCgesvd_bufferSize",
-        "cusolverDnZgesvd_bufferSize", "cusolverDnSgesvd", "cusolverDnDgesvd",
-        "cusolverDnCgesvd", "cusolverDnZgesvd", "cusolverDnSpotrfBatched",
-        "cusolverDnDpotrfBatched", "cusolverDnCpotrfBatched",
-        "cusolverDnZpotrfBatched", "cusolverDnSpotrsBatched",
-        "cusolverDnDpotrsBatched", "cusolverDnCpotrsBatched",
-        "cusolverDnZpotrsBatched", "cusolverDnSsygvd", "cusolverDnDsygvd",
-        "cusolverDnSsygvd_bufferSize", "cusolverDnDsygvd_bufferSize",
-        "cusolverDnChegvd", "cusolverDnZhegvd", "cusolverDnChegvd_bufferSize",
-        "cusolverDnZhegvd_bufferSize", "cusolverDnXgetrf",
-        "cusolverDnXgetrf_bufferSize", "cusolverDnXgetrs", "cusolverDnXgeqrf",
-        "cusolverDnXgeqrf_bufferSize", "cusolverDnGetrf",
-        "cusolverDnGetrf_bufferSize", "cusolverDnGetrs", "cusolverDnGeqrf",
-        "cusolverDnGeqrf_bufferSize", "cusolverDnCreateGesvdjInfo",
-        "cusolverDnDestroyGesvdjInfo", "cusolverDnCreateSyevjInfo",
-        "cusolverDnDestroySyevjInfo", "cusolverDnSgesvdj_bufferSize",
-        "cusolverDnDgesvdj_bufferSize", "cusolverDnCgesvdj_bufferSize",
-        "cusolverDnZgesvdj_bufferSize", "cusolverDnXgesvd_bufferSize",
-        "cusolverDnGesvd_bufferSize", "cusolverDnSgesvdj", "cusolverDnDgesvdj",
-        "cusolverDnCgesvdj", "cusolverDnZgesvdj", "cusolverDnXgesvd",
-        "cusolverDnGesvd", "cusolverDnXpotrf_bufferSize",
-        "cusolverDnPotrf_bufferSize", "cusolverDnXpotrf", "cusolverDnPotrf",
-        "cusolverDnXpotrs", "cusolverDnPotrs", "cusolverDnSsyevdx",
-        "cusolverDnDsyevdx", "cusolverDnSsyevdx_bufferSize",
-        "cusolverDnDsyevdx_bufferSize", "cusolverDnCheevdx",
-        "cusolverDnZheevdx", "cusolverDnCheevdx_bufferSize",
-        "cusolverDnZheevdx_bufferSize", "cusolverDnSsygvdx",
-        "cusolverDnDsygvdx", "cusolverDnSsygvdx_bufferSize",
-        "cusolverDnDsygvdx_bufferSize", "cusolverDnChegvdx",
-        "cusolverDnZhegvdx", "cusolverDnChegvdx_bufferSize",
-        "cusolverDnZhegvdx_bufferSize", "cusolverDnSsygvj", "cusolverDnDsygvj",
-        "cusolverDnSsygvj_bufferSize", "cusolverDnDsygvj_bufferSize",
-        "cusolverDnChegvj", "cusolverDnZhegvj", "cusolverDnChegvj_bufferSize",
-        "cusolverDnZhegvj_bufferSize", "cusolverDnXsyevdx",
-        "cusolverDnXsyevdx_bufferSize", "cusolverDnSyevdx",
-        "cusolverDnSyevdx_bufferSize", "cusolverDnSsyevj", "cusolverDnDsyevj",
-        "cusolverDnSsyevj_bufferSize", "cusolverDnDsyevj_bufferSize",
-        "cusolverDnCheevj", "cusolverDnZheevj", "cusolverDnCheevj_bufferSize",
-        "cusolverDnZheevj_bufferSize", "cusolverDnXsyevd",
-        "cusolverDnXsyevd_bufferSize", "cusolverDnSyevd",
-        "cusolverDnSyevd_bufferSize", "cusolverDnXtrtri",
-        "cusolverDnXtrtri_bufferSize", "cusolverDnSsyevd_bufferSize",
-        "cusolverDnDsyevd_bufferSize", "cusolverDnCheevd_bufferSize",
-        "cusolverDnZheevd_bufferSize", "cusolverDnSsyevd", "cusolverDnDsyevd",
-        "cusolverDnCheevd", "cusolverDnZheevd");
-  };
-
-  MF.addMatcher(callExpr(allOf(callee(functionDecl(functionName())),
-                               hasAncestor(functionDecl(
-                                   anyOf(hasAttr(attr::CUDADevice),
-                                         hasAttr(attr::CUDAGlobal))))))
-                    .bind("kernelCall"),
-                this);
-
-  MF.addMatcher(
-      callExpr(
-          allOf(callee(functionDecl(functionName())), parentStmt(),
-                hasAncestor(functionDecl(unless(allOf(
-                    hasAttr(attr::CUDADevice), hasAttr(attr::CUDAGlobal)))))))
-          .bind("FunctionCall"),
-      this);
-  MF.addMatcher(
-      callExpr(
-          allOf(callee(functionDecl(functionName())), unless(parentStmt()),
-                unless(hasParent(varDecl())),
-                hasAncestor(functionDecl(unless(allOf(
-                    hasAttr(attr::CUDADevice), hasAttr(attr::CUDAGlobal)))))))
-          .bind("FunctionCallUsedNotInitializeVarDecl"),
-      this);
-
-  MF.addMatcher(
-      callExpr(
-          allOf(callee(functionDecl(functionName())), hasParent(varDecl()),
-                hasAncestor(functionDecl(unless(allOf(
-                    hasAttr(attr::CUDADevice), hasAttr(attr::CUDAGlobal)))))))
-          .bind("FunctionCallUsedInitializeVarDecl"),
-      this);
-}
-
-void SOLVERFunctionCallRule::runRule(const MatchFinder::MatchResult &Result) {
-  bool IsAssigned = false;
-  bool IsInitializeVarDecl = false;
-  bool HasDeviceAttr = false;
-  const CallExpr *CE = getNodeAsType<CallExpr>(Result, "kernelCall");
-  if (CE) {
-    HasDeviceAttr = true;
-  } else if (!(CE = getNodeAsType<CallExpr>(Result, "FunctionCall"))) {
-    if ((CE = getNodeAsType<CallExpr>(
-             Result, "FunctionCallUsedNotInitializeVarDecl"))) {
-      IsAssigned = true;
-    } else if ((CE = getNodeAsType<CallExpr>(
-                    Result, "FunctionCallUsedInitializeVarDecl"))) {
-      IsAssigned = true;
-      IsInitializeVarDecl = true;
-    } else {
-      return;
-    }
-  }
-
-  const SourceManager *SM = Result.SourceManager;
-  auto SL = SM->getExpansionLoc(CE->getBeginLoc());
-  std::string Key = SM->getFilename(SL).str() +
-                    std::to_string(SM->getDecomposedLoc(SL).second);
-  DpctGlobalInfo::updateInitSuffixIndexInRule(
-      DpctGlobalInfo::getSuffixIndexInitValue(Key));
-
-  // Collect sourceLocations of the function call
-  SourceLocation FuncNameBegin(CE->getBeginLoc());
-  SourceLocation FuncCallEnd(CE->getEndLoc());
-
-  // Correct sourceLocations for macros
-  if (FuncNameBegin.isMacroID())
-    FuncNameBegin = SM->getExpansionLoc(FuncNameBegin);
-  if (FuncCallEnd.isMacroID())
-    FuncCallEnd = SM->getExpansionLoc(FuncCallEnd);
-  // Offset 1 is the length of the last token ")"
-  FuncCallEnd = FuncCallEnd.getLocWithOffset(1);
-
-  // Collect sourceLocations for creating new scope
-  std::string PrefixBeforeScope, PrefixInsertStr, SuffixInsertStr;
-  auto SR = getScopeInsertRange(CE, FuncNameBegin, FuncCallEnd);
-  SourceLocation StmtBegin = SR.getBegin(), StmtEndAfterSemi = SR.getEnd();
-  std::string IndentStr =
-      getIndent(StmtBegin, (Result.Context)->getSourceManager()).str();
-  Token Tok;
-  Lexer::getRawToken(FuncNameBegin, Tok, *SM, LangOptions());
-  SourceLocation FuncNameEnd = Tok.getEndLoc();
-  auto FuncNameLength =
-      SM->getCharacterData(FuncNameEnd) - SM->getCharacterData(FuncNameBegin);
-
-  // Prepare the prefix and the postfix for assignment
-  if (!CE->getDirectCallee())
-    return;
-  std::string FuncName =
-      CE->getDirectCallee()->getNameInfo().getName().getAsString();
-  std::string AssignPrefix = "";
-  std::string AssignPostfix = "";
-
-  if (IsAssigned) {
-    AssignPrefix = MapNames::getCheckErrorMacroName() + "(";
-    AssignPostfix = ")";
-  }
-
-  if (HasDeviceAttr) {
-    report(CE->getBeginLoc(), Diagnostics::FUNCTION_CALL_IN_DEVICE, false,
-           MapNames::ITFName.at(FuncName),
-           MemoryMigrationRule::getMemoryHelperFunctionName("memcpy"));
-    return;
-  }
-
-  const VarDecl *VD = 0;
-  if (IsInitializeVarDecl) {
-    // Create the prefix for VarDecl before scope and remove the VarDecl inside
-    // the scope
-    VD = getAncestralVarDecl(CE);
-    std::string VarType, VarName;
-    if (VD) {
-      VarType = VD->getType().getAsString();
-      VarName = VD->getNameAsString();
-
-      requestHelperFeatureForTypeNames(VarType);
-      insertHeaderForTypeRule(VarType, VD->getBeginLoc());
-      auto Itr = MapNames::TypeNamesMap.find(VarType);
-      if (Itr == MapNames::TypeNamesMap.end())
-        return;
-      VarType = Itr->second->NewName;
-      PrefixBeforeScope = VarType + " " + VarName + ";" + getNL() + IndentStr +
-                          PrefixBeforeScope;
-      SourceLocation typeBegin =
-          VD->getTypeSourceInfo()->getTypeLoc().getBeginLoc();
-      SourceLocation nameBegin = VD->getLocation();
-      SourceLocation nameEnd = Lexer::getLocForEndOfToken(
-          nameBegin, 0, *SM, Result.Context->getLangOpts());
-      auto replLen =
-          SM->getCharacterData(nameEnd) - SM->getCharacterData(typeBegin);
-      for (auto ItHeader = Itr->second->Includes.begin();
-           ItHeader != Itr->second->Includes.end(); ItHeader++) {
-        DpctGlobalInfo::getInstance().insertHeader(typeBegin, *ItHeader);
-      }
-      emplaceTransformation(
-          new ReplaceText(typeBegin, replLen, std::move(VarName)));
-    } else {
-      assert(0 && "Fail to get VarDecl information");
-      return;
-    }
-  }
-
-  if (MapNames::SOLVERAPIWithRewriter.find(FuncName) !=
-      MapNames::SOLVERAPIWithRewriter.end()) {
-    ExprAnalysis EA(CE);
-    emplaceTransformation(EA.getReplacement());
-    EA.applyAllSubExprRepl();
-    return;
-  } else if (MapNames::SOLVERFuncReplInfoMap.find(FuncName) !=
-      MapNames::SOLVERFuncReplInfoMap.end()) {
-    // Find replacement string
-    auto ReplInfoPair = MapNames::SOLVERFuncReplInfoMap.find(FuncName);
-    MapNames::SOLVERFuncReplInfo ReplInfo = ReplInfoPair->second;
-    std::string Replacement = ReplInfo.ReplName;
-
-    // Migrate arguments one by one
-    int ArgNum = CE->getNumArgs();
-    for (int i = 0; i < ArgNum; ++i) {
-      int IndexTemp = -1;
-      // MyFunction(float* a);
-      // In usm: Keep it and type cast if needed.
-      // In non-usm: Create a buffer, pass the buffer
-      // If type is int: MKL takes int64_t, so need to create a temp
-      //                 buffer/variable and copy the value back
-      //                 after the function call.
-      // Some API migration requires MoveFrom and MoveTo.
-      // e.g., move arg#1 to arg#0
-      // MyFunction(float* a, float* b);
-      // ---> MyFunction(float* b, float*a);
-      if (isReplIndex(i, ReplInfo.BufferIndexInfo, IndexTemp)) {
-        if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_None) {
-          requestFeature(HelperFeatureEnum::device_ext);
-          std::string BufferDecl;
-          std::string BufferName = getBufferNameAndDeclStr(
-              CE->getArg(i), *(Result.Context),
-              ReplInfo.BufferTypeInfo[IndexTemp], StmtBegin, BufferDecl, i);
-          PrefixInsertStr = PrefixInsertStr + BufferDecl;
-          if (ReplInfo.BufferTypeInfo[IndexTemp] == "int") {
-            PrefixInsertStr =
-                PrefixInsertStr + IndentStr + MapNames::getClNamespace() +
-                "buffer<int64_t> "
-                "result_temp_buffer" +
-                std::to_string(i) + "(" + MapNames::getClNamespace() +
-                "range<1>(1));" + getNL();
-            SuffixInsertStr = SuffixInsertStr + BufferName + ".get_access<" +
-                              MapNames::getClNamespace() +
-                              "access_mode::write>()[0] = "
-                              "(int)result_temp_buffer" +
-                              std::to_string(i) + ".get_access<" +
-                              MapNames::getClNamespace() +
-                              "access_mode::read>()[0];" + getNL() + IndentStr;
-            BufferName = "result_temp_buffer" + std::to_string(i);
-          }
-          bool Moved = false;
-          for (size_t j = 0; j < ReplInfo.MoveFrom.size(); j++) {
-            if (ReplInfo.MoveFrom[j] == i) {
-              Moved = true;
-              if (CE->getArg(ReplInfo.MoveTo[j])) {
-                emplaceTransformation(new InsertAfterStmt(
-                    CE->getArg(ReplInfo.MoveTo[j] - 1),
-                    ", result_temp_buffer" + std::to_string(i)));
-              }
-              ReplInfo.RedundantIndexInfo.push_back(i);
-              break;
-            }
-          }
-          if (!Moved) {
-            emplaceTransformation(new ReplaceStmt(CE->getArg(i), BufferName));
-          }
-        } else {
-          std::string ArgName = ExprAnalysis::ref(CE->getArg(i));
-          if (ReplInfo.BufferTypeInfo[IndexTemp] == "int") {
-            PrefixInsertStr = IndentStr + "int64_t result_temp_pointer" +
-                              std::to_string(i) + ";" + getNL();
-            SuffixInsertStr = SuffixInsertStr + " *" +
-                              ExprAnalysis::ref(CE->getArg(i)) +
-                              " = result_temp_pointer" + std::to_string(i) +
-                              ";" + getNL() + IndentStr;
-            ArgName = "&result_temp_pointer" + std::to_string(i);
-          }
-          bool Moved = false;
-          for (size_t j = 0; j < ReplInfo.MoveFrom.size(); j++) {
-            if (ReplInfo.MoveFrom[j] == i) {
-              Moved = true;
-              if (CE->getArg(ReplInfo.MoveTo[j])) {
-                emplaceTransformation(new InsertAfterStmt(
-                    CE->getArg(ReplInfo.MoveTo[j] - 1), ", " + ArgName));
-              }
-              ReplInfo.RedundantIndexInfo.push_back(i);
-              break;
-            }
-          }
-          if (!Moved) {
-            std::string TypeStr =
-                ReplInfo.BufferTypeInfo[IndexTemp].compare("int")
-                    ? "(" + ReplInfo.BufferTypeInfo[IndexTemp] + "*)"
-                    : "";
-            emplaceTransformation(
-                new ReplaceStmt(CE->getArg(i), TypeStr + ArgName));
-          }
-        }
-      }
-      // Remove the redundant args including the leading ","
-      if (isReplIndex(i, ReplInfo.RedundantIndexInfo, IndexTemp)) {
-        SourceLocation RemoveBegin, RemoveEnd;
-        if (i == 0) {
-          RemoveBegin = CE->getArg(i)->getBeginLoc();
-        } else {
-          RemoveBegin = CE->getArg(i - 1)->getEndLoc().getLocWithOffset(
-              Lexer::MeasureTokenLength(
-                  CE->getArg(i - 1)->getEndLoc(), *SM,
-                  dpct::DpctGlobalInfo::getContext().getLangOpts()));
-        }
-        RemoveEnd = CE->getArg(i)->getEndLoc().getLocWithOffset(
-            Lexer::MeasureTokenLength(
-                CE->getArg(i)->getEndLoc(), *SM,
-                dpct::DpctGlobalInfo::getContext().getLangOpts()));
-        auto ParameterLength =
-            SM->getCharacterData(RemoveEnd) - SM->getCharacterData(RemoveBegin);
-        emplaceTransformation(
-            new ReplaceText(RemoveBegin, ParameterLength, ""));
-      }
-      // OldFoo(float* out); --> *(out) = NewFoo();
-      // In current case, return value is always the last arg
-      if (ReplInfo.ReturnValue && i == ArgNum - 1) {
-        Replacement = "*(" +
-                      ExprAnalysis::ref(CE->getArg(CE->getNumArgs() - 1)) +
-                      ") = " + Replacement;
-      }
-      // The arg#0 is always the handler and will always be migrated to queue.
-      if (i == 0) {
-        // process handle argument
-        emplaceTransformation(new ReplaceStmt(
-            CE->getArg(i), "*" + ExprAnalysis::ref(CE->getArg(i))));
-      }
-    }
-    // Declare new args if it is used in MKL
-    if (!ReplInfo.MissedArgumentFinalLocation.empty()) {
-      std::string ReplStr;
-      for (size_t i = 0; i < ReplInfo.MissedArgumentFinalLocation.size(); ++i) {
-        if (ReplInfo.MissedArgumentIsBuffer[i]) {
-          PrefixInsertStr =
-              PrefixInsertStr + IndentStr + MapNames::getClNamespace() +
-              "buffer<" + ReplInfo.MissedArgumentType[i] + "> " +
-              ReplInfo.MissedArgumentName[i] + "(" +
-              MapNames::getClNamespace() + "range<1>(1));" + getNL();
-        } else {
-          PrefixInsertStr = PrefixInsertStr + IndentStr +
-                            ReplInfo.MissedArgumentType[i] + " " +
-                            ReplInfo.MissedArgumentName[i] + ";" + getNL();
-        }
-        ReplStr = ReplStr + ", " + ReplInfo.MissedArgumentName[i];
-        if (i == ReplInfo.MissedArgumentFinalLocation.size() - 1 ||
-            ReplInfo.MissedArgumentInsertBefore[i + 1] !=
-                ReplInfo.MissedArgumentInsertBefore[i]) {
-          if (ReplInfo.MissedArgumentInsertBefore[i] > 0) {
-            emplaceTransformation(new InsertAfterStmt(
-                CE->getArg(ReplInfo.MissedArgumentInsertBefore[i] - 1),
-                std::move(ReplStr)));
-          }
-          ReplStr = "";
-        }
-      }
-    }
-
-    // Copy an arg. e.g. copy arg#0 to arg#2
-    // OldFoo(int m, int n); --> NewFoo(int m, int n, int m)
-    if (!ReplInfo.CopyFrom.empty()) {
-      std::string InsStr = "";
-      for (size_t i = 0; i < ReplInfo.CopyFrom.size(); ++i) {
-        InsStr =
-            InsStr + ", " + ExprAnalysis::ref(CE->getArg(ReplInfo.CopyFrom[i]));
-        if (i == ReplInfo.CopyTo.size() - 1 ||
-            ReplInfo.CopyTo[i + 1] != ReplInfo.CopyTo[i]) {
-          emplaceTransformation(new InsertAfterStmt(
-              CE->getArg(ReplInfo.CopyTo[i - 1]), std::move(InsStr)));
-          InsStr = "";
-        }
-      }
-    }
-    // Type cast, for enum type migration
-    if (!ReplInfo.CastIndexInfo.empty()) {
-      for (size_t i = 0; i < ReplInfo.CastIndexInfo.size(); ++i) {
-        std::string CastStr = "(" + ReplInfo.CastTypeInfo[i] + ")";
-        emplaceTransformation(new InsertBeforeStmt(
-            CE->getArg(ReplInfo.CastIndexInfo[i]), std::move(CastStr)));
-      }
-    }
-    // Create scratchpad and scratchpad_size if required in MKL
-    if (!ReplInfo.WorkspaceIndexInfo.empty()) {
-      std::string BufferSizeArgStr = "";
-      for (size_t i = 0; i < ReplInfo.WorkspaceSizeInfo.size(); ++i) {
-        BufferSizeArgStr += i ? " ," : "";
-        BufferSizeArgStr +=
-            ExprAnalysis::ref(CE->getArg(ReplInfo.WorkspaceSizeInfo[i]));
-      }
-      std::string ScratchpadSizeNameStr =
-          "scratchpad_size_ct" +
-          std::to_string(dpct::DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-      std::string ScratchpadNameStr =
-          "scratchpad_ct" +
-          std::to_string(dpct::DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-      std::string EventNameStr =
-          "event_ct" +
-          std::to_string(dpct::DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-      std::string WSVectorNameStr =
-          "ws_vec_ct" +
-          std::to_string(dpct::DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-      PrefixInsertStr += IndentStr + "std::int64_t " + ScratchpadSizeNameStr +
-                         " = " + ReplInfo.WorkspaceSizeFuncName + "(*" +
-                         BufferSizeArgStr + ");" + getNL();
-      std::string BufferTypeStr = "float";
-      if (ReplInfo.BufferTypeInfo.size() > 0) {
-        BufferTypeStr = ReplInfo.BufferTypeInfo[0];
-      }
-      if (DpctGlobalInfo::getUsmLevel() == UsmLevel::UL_Restricted) {
-        DpctGlobalInfo::getInstance().insertHeader(CE->getBeginLoc(),
-                                                   HT_Thread);
-
-        PrefixInsertStr +=
-            IndentStr + BufferTypeStr + " *" + ScratchpadNameStr + " = " +
-            MapNames::getClNamespace() + "malloc_device<" + BufferTypeStr +
-            ">(" + ScratchpadSizeNameStr + ", *" +
-            ExprAnalysis::ref(CE->getArg(0)) + ");" + getNL();
-        PrefixInsertStr += IndentStr + MapNames::getClNamespace() + "event " +
-                           EventNameStr + ";" + getNL();
-
-        Replacement = EventNameStr + " = " + Replacement;
-
-        SuffixInsertStr += "std::vector<void *> " + WSVectorNameStr + "{" +
-                           ScratchpadNameStr + "};" + getNL() + IndentStr;
-        SuffixInsertStr +=
-            MapNames::getDpctNamespace() +
-            (DpctGlobalInfo::useSYCLCompat() ? "enqueue_free("
-                                             : "async_dpct_free(") +
-            WSVectorNameStr + ", {" + EventNameStr + "}, *" +
-            ExprAnalysis::ref(CE->getArg(0)) + ");" + getNL() + IndentStr;
-        requestFeature(HelperFeatureEnum::device_ext);
-      } else {
-        PrefixInsertStr += IndentStr + MapNames::getClNamespace() + "buffer<" +
-                           BufferTypeStr + ", 1> " + ScratchpadNameStr + "{" +
-                           MapNames::getClNamespace() + "range<1>(" +
-                           ScratchpadSizeNameStr + ")};" + getNL();
-      }
-      if (ReplInfo.WorkspaceIndexInfo[0] > 0) {
-        emplaceTransformation(new InsertAfterStmt(
-            CE->getArg(ReplInfo.WorkspaceIndexInfo[0]),
-            ", " + ScratchpadNameStr + ", " + ScratchpadSizeNameStr));
-      }
-    }
-
-    // Create scratchpad_size if only scratchpad_size is required in MKL
-    if (!ReplInfo.WSSizeInsertAfter.empty()) {
-      std::string BufferSizeArgStr = "";
-      for (size_t i = 0; i < ReplInfo.WSSizeInfo.size(); ++i) {
-        BufferSizeArgStr += i ? " ," : "";
-        BufferSizeArgStr +=
-            ExprAnalysis::ref(CE->getArg(ReplInfo.WSSizeInfo[i]));
-      }
-      std::string ScratchpadSizeNameStr =
-          "scratchpad_size_ct" +
-          std::to_string(dpct::DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-      PrefixInsertStr += IndentStr + "std::int64_t " + ScratchpadSizeNameStr +
-                         " = " + ReplInfo.WSSFuncName + "(*" +
-                         BufferSizeArgStr + ");" + getNL();
-      if (ReplInfo.WSSizeInsertAfter[0] > 0) {
-        emplaceTransformation(
-            new InsertAfterStmt(CE->getArg(ReplInfo.WSSizeInsertAfter[0]),
-                                ", " + ScratchpadSizeNameStr));
-      }
-    }
-
-    // Check PrefixInsertStr and SuffixInsertStr to decide whether to add
-    // bracket
-    std::string PrefixWithBracket = "";
-    std::string SuffixWithBracket = "";
-    if (!PrefixInsertStr.empty() || !SuffixInsertStr.empty()) {
-      PrefixWithBracket =
-          "{" + std::string(getNL()) + PrefixInsertStr + IndentStr;
-      SuffixWithBracket = getNL() + IndentStr + SuffixInsertStr + "}";
-    }
-
-    std::string ReplaceFuncName = Replacement;
-    emplaceTransformation(
-        new ReplaceText(FuncNameBegin, FuncNameLength, std::move(Replacement)));
-    insertAroundRange(StmtBegin, StmtEndAfterSemi,
-                      PrefixBeforeScope + PrefixWithBracket,
-                      std::move(SuffixWithBracket));
-
-    StringRef FuncNameRef(FuncName);
-    if (FuncNameRef.ends_with("getrf")) {
-      report(StmtBegin, Diagnostics::DIFFERENT_LU_FACTORIZATION, true,
-             getStmtSpelling(CE->getArg(6)), ReplaceFuncName,
-             MapNames::ITFName.at(FuncName));
-    }
-    if (IsAssigned) {
-      insertAroundRange(FuncNameBegin, FuncCallEnd, std::move(AssignPrefix),
-                        std::move(AssignPostfix));
-      requestFeature(HelperFeatureEnum::device_ext);
-    }
-  } else if (FuncName == "cusolverDnCreate" ||
-             FuncName == "cusolverDnDestroy") {
-    std::string Repl;
-    if (FuncName == "cusolverDnCreate") {
-      std::string LHS = getDrefName(CE->getArg(0));
-      if (isPlaceholderIdxDuplicated(CE))
-        return;
-      int Index = DpctGlobalInfo::getHelperFuncReplInfoIndexThenInc();
-      buildTempVariableMap(Index, CE, HelperFuncType::HFT_DefaultQueuePtr);
-      Repl = LHS + " = {{NEEDREPLACEZ" + std::to_string(Index) + "}}";
-    } else if (FuncName == "cusolverDnDestroy") {
-      dpct::ExprAnalysis EA(CE->getArg(0));
-      Repl = EA.getReplacedString() + " = nullptr";
-    } else {
-      return;
-    }
-
-    if (IsAssigned) {
-      requestFeature(HelperFeatureEnum::device_ext);
-      emplaceTransformation(
-          new ReplaceStmt(CE, true, MapNames::getCheckErrorMacroName() + "(" + Repl + ")"));
-    } else {
-      emplaceTransformation(new ReplaceStmt(CE, true, Repl));
-    }
-  }
-}
-
-void SOLVERFunctionCallRule::getParameterEnd(
-    const SourceLocation &ParameterEnd, SourceLocation &ParameterEndAfterComma,
-    const ast_matchers::MatchFinder::MatchResult &Result) {
-  std::optional<Token> TokSharedPtr;
-  TokSharedPtr = Lexer::findNextToken(ParameterEnd, *(Result.SourceManager),
-                                      LangOptions());
-  Token TokComma = TokSharedPtr.value();
-  if (TokComma.getKind() == tok::comma) {
-    ParameterEndAfterComma = TokComma.getEndLoc();
-  } else {
-    ParameterEndAfterComma = TokComma.getLocation();
-  }
-}
-
-bool SOLVERFunctionCallRule::isReplIndex(int Input, std::vector<int> &IndexInfo,
-                                         int &IndexTemp) {
-  for (int i = 0; i < static_cast<int>(IndexInfo.size()); ++i) {
-    if (IndexInfo[i] == Input) {
-      IndexTemp = i;
-      return true;
-    }
-  }
-  return false;
-}
-
-std::string SOLVERFunctionCallRule::getBufferNameAndDeclStr(
-    const Expr *Arg, const ASTContext &AC, const std::string &TypeAsStr,
-    SourceLocation SL, std::string &BufferDecl, int DistinctionID) {
-
-  std::string PointerName = ExprAnalysis::ref(Arg);
-  std::string BufferTempName =
-      getTempNameForExpr(Arg, true, true) + "buf_ct" +
-      std::to_string(dpct::DpctGlobalInfo::getSuffixIndexInRuleThenInc());
-
-  requestFeature(HelperFeatureEnum::device_ext);
-  BufferDecl = getIndent(SL, AC.getSourceManager()).str() + "auto " +
-               BufferTempName + " = " + MapNames::getDpctNamespace() +
-               "get_buffer<" + TypeAsStr + ">(" + PointerName + ");" + getNL();
-  return BufferTempName;
-}
-
-const clang::VarDecl *
-SOLVERFunctionCallRule::getAncestralVarDecl(const clang::CallExpr *CE) {
-  auto &Context = dpct::DpctGlobalInfo::getContext();
-  auto Parents = Context.getParents(*CE);
-  while (Parents.size() == 1) {
-    auto *Parent = Parents[0].get<VarDecl>();
-    if (Parent) {
-      return Parent;
-    } else {
-      Parents = Context.getParents(Parents[0]);
-    }
-  }
-  return nullptr;
-}
 
 void FunctionCallRule::registerMatcher(MatchFinder &MF) {
   auto functionName = [&]() {
